Broad sense heritability calculation.

This dataset contains phenotpic data for 41 potato lines evaluated in 3 environments in an RCBD design. The phenotypic trait is tuber quality and we show how to obtain an estimate of DT_example for the trait.

Usage

data("DT_example")

Format

The format is: chr "DT_example"

Source

This data was generated by a potato study.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

See also

Examples

####=========================================####
#### For CRAN time limitations most lines in the 
#### examples are silenced with one '#' mark, 
#### remove them and run the examples
####=========================================####
####=========================================####
#### EXAMPLES
#### Different models with sommer
####=========================================####

data(DT_example)
DT <- DT_example
A <- A_example
head(DT)
#>                   Name     Env Loc Year     Block Yield    Weight
#> 33  Manistee(MSL292-A) CA.2013  CA 2013 CA.2013.1     4 -1.904711
#> 65          CO02024-9W CA.2013  CA 2013 CA.2013.1     5 -1.446958
#> 66  Manistee(MSL292-A) CA.2013  CA 2013 CA.2013.2     5 -1.516271
#> 67            MSL007-B CA.2011  CA 2011 CA.2011.2     5 -1.435510
#> 68           MSR169-8Y CA.2013  CA 2013 CA.2013.1     5 -1.469051
#> 103         AC05153-1W CA.2013  CA 2013 CA.2013.1     6 -1.307167

####=========================================####
#### Univariate homogeneous variance models  ####
####=========================================####

## Compound simmetry (CS) model
ans1 <- mmer(Yield~Env,
             random= ~ Name + Env:Name,
             rcov= ~ units,
             data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -31.2668   1:14:59      0           0
#>     2      -23.2804   1:14:59      0           0
#>     3      -20.4746   1:14:59      0           0
#>     4      -20.1501   1:14:59      0           0
#>     5      -20.1454   1:14:59      0           0
#>     6      -20.1454   1:14:59      0           0
summary(ans1)$varcomp
#>                       VarComp VarCompSE   Zratio Constraint
#> Name.Yield-Yield     3.681877 1.6909561 2.177394   Positive
#> Env:Name.Yield-Yield 5.173062 1.4952313 3.459707   Positive
#> units.Yield-Yield    4.366285 0.6470458 6.748031   Positive

ans1b <- mmec(Yield~Env,
             random= ~ Name + Env:Name,
             rcov= ~ units,
             data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -327.295   1:15:0      0           0
#>     2      -311.454   1:15:0      0           0
#>     3      -304.865   1:15:0      0           0
#>     4      -302.661   1:15:0      0           0
#>     5      -301.766   1:15:0      0           0
#>     6      -301.251   1:15:0      0           0
#>     7      -300.978   1:15:0      0           0
#>     8      -300.877   1:15:0      0           0
#>     9      -300.855   1:15:0      0           0
#>     10      -300.852   1:15:0      0           0
#>     11      -300.852   1:15:0      0           0
summary(ans1b)$varcomp
#>                   VarComp VarCompSE   Zratio Constraint
#> Name:isc:isc     3.683153  1.880974 1.958109   Positive
#> Env:Name:isc:isc 5.174116  2.420486 2.137635   Positive
#> units:isc:isc    4.359651  2.270467 1.920156   Positive

# ####===========================================####
# #### Univariate heterogeneous variance models  ####
# ####===========================================####
# 
# ## Compound simmetry (CS) + Diagonal (DIAG) model
# ans2 <- mmer(Yield~Env,
#              random= ~Name + vsr(dsr(Env),Name),
#              rcov= ~ vsr(dsr(Env),units),
#              data=DT)
# summary(ans2)
# 
# DT=DT[with(DT, order(Env)), ]
# ans2b <- mmec(Yield~Env,
#              random= ~Name + vsc(dsc(Env),isc(Name)) +
#                       vsc(atc(Env, c("CA.2011") ),isc(Block)) ,
#              rcov= ~ vsc(dsc(Env),isc(units)),
#              data=DT)
# summary(ans2b)
# 
# ####==========================================####
# #### Multivariate homogeneous variance models ####
# ####==========================================####
# 
# ## Multivariate Compound simmetry (CS) model
# DT$EnvName <- paste(DT$Env,DT$Name)
# ans4 <- mmer(cbind(Yield, Weight) ~ Env,
#               random= ~ vsr(Name) + vsr(EnvName),
#               rcov= ~ vsr(units),
#               data=DT)
# summary(ans4)