Full diallel data for corn hybrids

This dataset contains phenotpic data for 36 winter bean hybrids, coming from a full diallel design and evaluated for 9 traits. The column male and female origin columns are included as well.

Usage

data("DT_fulldiallel")

Format

The format is: chr "DT_fulldiallel"

Source

This data was generated by a winter bean study and originally included in the agridat package.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

See also

Examples

####=========================================####
#### For CRAN time limitations most lines in the
#### examples are silenced with one '#' mark,
#### remove them and run the examples
####=========================================####
data(DT_fulldiallel)
DT <- DT_fulldiallel
head(DT)
#>   female male yield stems nodes pods seeds weight height width flower
#> 1    G24  G24 172.8  20.5  2.99 1.45  2.91   68.7  20.05 16.04  10.38
#> 2    G31  G24 247.7  26.8  3.24 1.39  3.00   69.5  25.13 22.29  10.68
#> 3    G36  G24 267.5  26.3  3.42 1.31  3.13   75.5  23.20 20.88   7.93
#> 4    G55  G24 301.7  28.8  3.98 1.34  2.64   77.2  25.20 24.00   5.37
#> 5    G64  G24 267.5  29.7  3.34 1.27  3.15   69.8  23.97 23.67   3.35
#> 6    G67  G24 262.0  28.3  2.69 1.52  2.80   81.8  23.33 20.79  12.00
mix <- mmer(stems~1, random=~female+male, data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -15.1177   1:15:0      0           0
#>     2      -14.9349   1:15:0      0           0
#>     3      -14.8855   1:15:0      0           0
#>     4      -14.8811   1:15:0      0           0
#>     5      -14.881   1:15:0      0           0
summary(mix)
#> ============================================================
#>          Multivariate Linear Mixed Model fit by REML         
#> **********************  sommer 4.3  ********************** 
#> ============================================================
#>          logLik      AIC      BIC Method Converge
#> Value -14.88101 31.76201 33.34553     NR     TRUE
#> ============================================================
#> Variance-Covariance components:
#>                    VarComp VarCompSE Zratio Constraint
#> female.stems-stems   1.984     1.829  1.085   Positive
#> male.stems-stems     1.775     1.699  1.045   Positive
#> units.stems-stems    5.292     1.497  3.535   Positive
#> ============================================================
#> Fixed effects:
#>   Trait      Effect Estimate Std.Error t.value
#> 1 stems (Intercept)    27.52    0.8795   31.29
#> ============================================================
#> Groups and observations:
#>        stems
#> female     6
#> male       6
#> ============================================================
#> Use the '$' sign to access results and parameters

mixb <- mmec(stems~1, random=~female+male, data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -54.4362   1:15:0      0           0
#>     2      -54.2835   1:15:0      0           0
#>     3      -54.2126   1:15:0      0           0
#>     4      -54.176   1:15:0      0           0
#>     5      -54.1604   1:15:0      0           0
#>     6      -54.1551   1:15:0      0           0
#>     7      -54.1536   1:15:0      0           0
#>     8      -54.1532   1:15:0      0           0
summary(mixb)$varcomp
#>                 VarComp VarCompSE   Zratio Constraint
#> female:isc:isc 1.974269  1.505941 1.310987   Positive
#> male:isc:isc   1.769532  1.507346 1.173939   Positive
#> units:isc:isc  5.292799  1.806661 2.929603   Positive

####=========================================####
####=========================================####
#### Multivariate model example
####=========================================####
####=========================================####

data(DT_fulldiallel)
DT <- DT_fulldiallel
head(DT)
#>   female male yield stems nodes pods seeds weight height width flower
#> 1    G24  G24 172.8  20.5  2.99 1.45  2.91   68.7  20.05 16.04  10.38
#> 2    G31  G24 247.7  26.8  3.24 1.39  3.00   69.5  25.13 22.29  10.68
#> 3    G36  G24 267.5  26.3  3.42 1.31  3.13   75.5  23.20 20.88   7.93
#> 4    G55  G24 301.7  28.8  3.98 1.34  2.64   77.2  25.20 24.00   5.37
#> 5    G64  G24 267.5  29.7  3.34 1.27  3.15   69.8  23.97 23.67   3.35
#> 6    G67  G24 262.0  28.3  2.69 1.52  2.80   81.8  23.33 20.79  12.00

mix <- mmer(cbind(stems,pods,seeds)~1,
             random=~vsr(female) + vsr(male),
             rcov=~vsr(units),
             data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -63.6729   1:15:1      1           0
#>     2      -40.8016   1:15:1      1           0
#>     3      -27.2147   1:15:1      1           0
#>     4      -24.2403   1:15:1      1           0
#>     5      -24.1945   1:15:1      1           0
#>     6      -24.194   1:15:1      1           0
summary(mix)
#> ============================================================
#>          Multivariate Linear Mixed Model fit by REML         
#> **********************  sommer 4.3  ********************** 
#> ============================================================
#>          logLik      AIC      BIC Method Converge
#> Value -24.19402 54.38805 62.43444     NR     TRUE
#> ============================================================
#> Variance-Covariance components:
#>                        VarComp VarCompSE  Zratio Constraint
#> u:female.stems-stems  1.983920  1.829197  1.0846   Positive
#> u:female.stems-pods   0.047557  0.056891  0.8359   Unconstr
#> u:female.stems-seeds -0.202772  0.156026 -1.2996   Unconstr
#> u:female.pods-pods    0.003453  0.003268  1.0565   Positive
#> u:female.pods-seeds  -0.007714  0.006429 -1.1998   Unconstr
#> u:female.seeds-seeds  0.027291  0.018584  1.4686   Positive
#> u:male.stems-stems    1.775485  1.698746  1.0452   Positive
#> u:male.stems-pods     0.030613  0.061605  0.4969   Unconstr
#> u:male.stems-seeds   -0.171257  0.162701 -1.0526   Unconstr
#> u:male.pods-pods      0.005240  0.004390  1.1938   Positive
#> u:male.pods-seeds    -0.006019  0.007864 -0.7654   Unconstr
#> u:male.seeds-seeds    0.037802  0.025226  1.4985   Positive
#> u:units.stems-stems   5.291650  1.496464  3.5361   Positive
#> u:units.stems-pods   -0.077067  0.048449 -1.5907   Unconstr
#> u:units.stems-seeds   0.059026  0.052802  1.1179   Unconstr
#> u:units.pods-pods     0.009972  0.002820  3.5360   Positive
#> u:units.pods-seeds    0.001446  0.002253  0.6417   Unconstr
#> u:units.seeds-seeds   0.012510  0.003540  3.5343   Positive
#> ============================================================
#> Fixed effects:
#>   Trait      Effect Estimate Std.Error t.value
#> 1 stems (Intercept)   27.519   0.87952   31.29
#> 2  pods (Intercept)    1.336   0.04154   32.17
#> 3 seeds (Intercept)    2.855   0.10581   26.98
#> ============================================================
#> Groups and observations:
#>          stems pods seeds
#> u:female     6    6     6
#> u:male       6    6     6
#> ============================================================
#> Use the '$' sign to access results and parameters
#### genetic variance covariance
cov2cor(mix$sigma$`u:female`)
#>            stems       pods      seeds
#> stems  1.0000000  0.5745963 -0.8714338
#> pods   0.5745963  1.0000000 -0.7946327
#> seeds -0.8714338 -0.7946327  1.0000000
cov2cor(mix$sigma$`u:male`)
#>            stems       pods      seeds
#> stems  1.0000000  0.3173646 -0.6610472
#> pods   0.3173646  1.0000000 -0.4276277
#> seeds -0.6610472 -0.4276277  1.0000000
cov2cor(mix$sigma$`u:units`)
#>            stems       pods     seeds
#> stems  1.0000000 -0.3354943 0.2294149
#> pods  -0.3354943  1.0000000 0.1294200
#> seeds  0.2294149  0.1294200 1.0000000