half diallel data for corn hybrids

This dataset contains phenotpic data for 21 corn hybrids, with 2 technical repetitions, coming from a half diallel design and evaluated for sugar content. The column geno indicates the hybrid and male and female origin columns are included as well.

Usage

data("DT_halfdiallel")

Format

The format is: chr "DT_halfdiallel"

Source

This data was generated by a corn study.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

See also

Examples

####=========================================####
#### For CRAN time limitations most lines in the 
#### examples are silenced with one '#' mark, 
#### remove them and run the examples
####=========================================####

data("DT_halfdiallel")
DT <- DT_halfdiallel
head(DT)
#>   rep geno male female     sugar
#> 1   1   12    1      2 13.950509
#> 2   2   12    1      2  9.756918
#> 3   1   13    1      3 13.906355
#> 4   2   13    1      3  9.119455
#> 5   1   14    1      4  5.174483
#> 6   2   14    1      4  8.452221
DT$femalef <- as.factor(DT$female)
DT$malef <- as.factor(DT$male)
DT$genof <- as.factor(DT$geno)

A <- diag(7); colnames(A) <- rownames(A) <- 1:7;A # if you want to provide a covariance matrix
#>   1 2 3 4 5 6 7
#> 1 1 0 0 0 0 0 0
#> 2 0 1 0 0 0 0 0
#> 3 0 0 1 0 0 0 0
#> 4 0 0 0 1 0 0 0
#> 5 0 0 0 0 1 0 0
#> 6 0 0 0 0 0 1 0
#> 7 0 0 0 0 0 0 1
#### model using overlay
modh <- mmer(sugar~1, 
             random=~vsr(overlay(femalef,malef, sparse = FALSE), Gu=A) 
             + genof,
             data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -10.425   1:15:1      0           0
#>     2      -6.487   1:15:1      0           0
#>     3      -5.732   1:15:1      0           0
#>     4      -5.67494   1:15:1      0           0
#>     5      -5.67441   1:15:1      0           0
summary(modh)$varcomp
#>                              VarComp VarCompSE   Zratio Constraint
#> u:femalef:malef.sugar-sugar 5.507899 3.5741151 1.541052   Positive
#> genof.sugar-sugar           1.815784 1.3629575 1.332238   Positive
#> units.sugar-sugar           3.117538 0.9626094 3.238632   Positive

Ai <- as(solve(A + diag(1e-4,ncol(A),ncol(A))), Class="dgCMatrix")
modhb <- mmec(sugar~1, 
             random=~vsc(isc(overlay(femalef,malef, sparse = TRUE)), Gu=Ai) 
             + genof,
             data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -66.9631   1:15:1      0           0
#>     2      -62.779   1:15:1      0           0
#>     3      -60.5877   1:15:1      0           0
#>     4      -60.1123   1:15:1      0           0
#>     5      -60.0669   1:15:1      0           0
#>     6      -60.0549   1:15:1      0           0
#>     7      -60.05   1:15:1      0           0
#>     8      -60.0481   1:15:1      0           0
#>     9      -60.0473   1:15:1      0           0
summary(modhb)$varcomp
#>                           VarComp VarCompSE   Zratio Constraint
#> femalef:malef:Ai:isc:isc 5.421024  1.956277 2.771092   Positive
#> genof:isc:isc            1.815385  2.972937 0.610637   Positive
#> units:isc:isc            3.119559  2.997912 1.040577   Positive