Reaction times in a sleep deprivation study

The average reaction time per day for subjects in a sleep deprivation study. On day 0 the subjects had their normal amount of sleep. Starting that night they were restricted to 3 hours of sleep per night. The observations represent the average reaction time on a series of tests given each day to each subject. Data from sleepstudy to see how lme4 models can be translated in sommer.

Usage

data("DT_sleepstudy")

Format

The format is: chr "DT_sleepstudy"

Source

These data are from the study described in Belenky et al. (2003), for the sleep deprived group and for the first 10 days of the study, up to the recovery period.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Gregory Belenky et al. (2003) Patterns of performance degradation and restoration during sleep restrictions and subsequent recovery: a sleep dose-response study. Journal of Sleep Research 12, 1-12.

See also

Examples

####=========================================####
#### For CRAN time limitations most lines in the
#### examples are silenced with one '#' mark,
#### remove them and run the examples
####=========================================####
# library(lme4)
data(DT_sleepstudy)
DT <- DT_sleepstudy
head(DT)
#>   Reaction Days Subject Subjectf Daysf
#> 1 249.5600    0     308      308     0
#> 2 258.7047    1     308      308     1
#> 3 250.8006    2     308      308     2
#> 4 321.4398    3     308      308     3
#> 5 356.8519    4     308      308     4
#> 6 414.6901    5     308      308     5
##################################
## lme4
# fm1 <- lmer(Reaction ~ Days + (1 | Subject), data=DT)
# vc <- VarCorr(fm1); print(vc,comp=c("Variance"))
## sommer
fm2 <- mmer(Reaction ~ Days,
            random= ~ Subject, 
            data=DT, tolParInv = 1e-6, verbose = FALSE)
summary(fm2)$varcomp
#>                             VarComp VarCompSE   Zratio Constraint
#> Subject.Reaction-Reaction 1377.9758  505.0776 2.728246   Positive
#> units.Reaction-Reaction    960.4705  107.0638 8.971013   Positive

##################################
## lme4
# fm1 <- lmer(Reaction ~ Days + (Days || Subject), data=DT)
# vc <- VarCorr(fm1); print(vc,comp=c("Variance"))
## sommer
fm2 <- mmer(Reaction ~ Days,
            random= ~ Subject + vsr(Days, Subject), 
            data=DT, tolParInv = 1e-6, verbose = FALSE)
summary(fm2)$varcomp
#>                                  VarComp VarCompSE   Zratio Constraint
#> Subject.Reaction-Reaction      627.54087 283.52939 2.213319   Positive
#> Days:Subject.Reaction-Reaction  35.86008  14.53187 2.467686   Positive
#> units.Reaction-Reaction        653.58305  76.72711 8.518281   Positive

##################################
## lme4
# fm1 <- lmer(Reaction ~ Days + (Days | Subject), data=DT)
# vc <- VarCorr(fm1); print(vc,comp=c("Variance"))
## sommer
## no equivalence in sommer to find the correlation between the 2 vc
## this is the most similar which is equivalent to (intercept || slope)
fm2 <- mmer(Reaction ~ Days,
            random= ~ Subject + vsr(Days, Subject), 
            data=DT, tolParInv = 1e-6, verbose = FALSE)
summary(fm2)$varcomp
#>                                  VarComp VarCompSE   Zratio Constraint
#> Subject.Reaction-Reaction      627.54087 283.52939 2.213319   Positive
#> Days:Subject.Reaction-Reaction  35.86008  14.53187 2.467686   Positive
#> units.Reaction-Reaction        653.58305  76.72711 8.518281   Positive

##################################
## lme4
# fm1 <- lmer(Reaction ~ Days + (0 + Days | Subject), data=DT)
# vc <- VarCorr(fm1); print(vc,comp=c("Variance"))
## sommer
fm2 <- mmer(Reaction ~ Days,
            random= ~ vsr(Days, Subject), 
            data=DT, tolParInv = 1e-6, verbose = FALSE)
summary(fm2)$varcomp
#>                                  VarComp VarCompSE   Zratio Constraint
#> Days:Subject.Reaction-Reaction  52.70946  19.09984 2.759681   Positive
#> units.Reaction-Reaction        842.02736  93.84640 8.972399   Positive