Yield of oats in a split-block experiment

The yield of oats from a split-plot field trial using three varieties and four levels of manurial treatment. The experiment was laid out in 6 blocks of 3 main plots, each split into 4 sub-plots. The varieties were applied to the main plots and the manurial (nitrogen) treatments to the sub-plots.

Format

block

block factor with 6 levels

nitro

nitrogen treatment in hundredweight per acre

Variety

genotype factor, 3 levels

yield

yield in 1/4 lbs per sub-plot, each 1/80 acre.

row

row location

column

column location

Source

Yates, Frank (1935) Complex experiments, Journal of the Royal Statistical Society Suppl. 2, 181–247.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

Examples

### ========================== ###
### ========================== ###
data(DT_yatesoats)
DT <- DT_yatesoats
head(DT)
#>   row col   Y   N          V  B         MP
#> 1   1   1  91 0.2    Victory B2    Victory
#> 2   2   1  61   0    Victory B2    Victory
#> 3   3   1 121 0.4 Marvellous B2 Marvellous
#> 4   4   1 144 0.6 Marvellous B2 Marvellous
#> 5   5   1 149 0.6 GoldenRain B2 GoldenRain
#> 6   6   1 108 0.2 GoldenRain B2 GoldenRain
# m3 <- mmer(fixed=Y ~ V + N + V:N,
#            random = ~ B + B:MP,
#            rcov=~units,
#            data = DT)
# summary(m3)$varcomp
# 
# m3b <- mmec(fixed=Y ~ V + N + V:N,
#            random = ~ B + B:MP,
#            rcov=~units,
#            data = DT)
# summary(m3b)$varcomp