Multivariate Newton-Raphson algorithm

Multivariate Newton-Raphson algorithm used behind mmes when the henderson argument is set to FALSE. Algorithm made available for users that want to avoid the user-friendly interface and provide their matrices directly.

Usage

MNR(Y, # response
    X, Gx, # fixed effects
    Z, K, # random effects
    R,  # residual effects
    Ge, GeI, # initial values and constraints
    W, isInvW, # weights matrix
    iters, tolpar, tolparinv, # other params
    ai, pev,
    verbose, retscaled,
    stepweight, emweight
    )

Arguments

Y

A matrix with rows for records and columns for traits. Expected class is a 'matrix'.

X

Design matrices for fixed effects. Expected class is a 'list' with as many design matrices as fixed effects. Matrices within the list should be of class 'matrix'

Gx

Trait multiplier matrix for fixed effects. Expected class is a 'list' with as many matrices as fixed effects. Each matrix is a square matrix with as many rows and columns as number of traits. Each matrix is of class 'matrix'

Z

Design matrices for random effects. Expected class is a 'list' with as many design matrices as random effects. Each element in the list should be a matrix of class 'dgCMatrix'

K

Covariance matrices for design matrices of random effects. Expected class is a 'list' with as many covariance matrices as random effects specified in Z. Each element in the list should be a matrix of class 'dgCMatrix'

R

Residual matrices for residual effects. Expected class is a 'list' with as many residual matrices as residual effects. Each element in the list should be a square matrix of dimensions n x n, where n is the number of records. Each matrix should be of class 'matrix'

Ge

Initial values for variance components. Expected class is a 'list' with as many variance component matrices as random effects specified in Z. Each element in the list should be a matrix of dimensions t x t, where t is the number of traits and of class 'matrix'. Initial values are any real values.

GeI

Initial constraints for variance components. Expected class is a 'list' with as many variance component constraints matrices as random effects specified in Z. Each element in the list should be a matrix of dimensions t x t, where t is the number of traits and of class 'matrix'. Values expected are:

0: not to be estimated

1: estimated and constrained to be positive (i.e. variance component)

2: estimated and unconstrained (can be negative or positive, i.e. covariance component)

3: not to be estimated but fixed (value has to be provided in the Gti argument)

Please notice that lower triangular values of these matrices have to be equal to zero since only the values in the upper triangular are estimated.

W

Design matrix for weights. A matrix of class 'matrix' for weighting the records.

isInvW

A value of class 'logical' to indicate if the W matrix provided is already an inverse or not. This aims to speed up the computations.

iters

Maximum number of iterations allowed in REML. Value is of class 'integer'.

tolpar

Convergence criteria for the change in log-likelihood. Value is of class 'numeric'.

tolparinv

Tolerance parameter for matrix inverse used when singularities are encountered in the estimation procedure. Value is of class 'numeric'.

ai

A value of class 'logical' to indicate if the Average Information algorithm should be used instead. That is faster but much less stable.

pev

A value of class 'logical' to indicate if the predicted error variance should be computed or not. If FALSE computations are speeded up for models with many effects to be estimated.

verbose

A value of class 'logical' to value to indicate if the program should return the progress of the iterative algorithm.

retscaled

A value of class 'logical' to indicate if we should avoid scaling the traits.

stepweight

Vector of class 'numeric' with length equal to niters to specify the relative weight given to the second derivative Newton update.

emweight

Vector of class 'numeric' with length equal to niters to specify the relative weight given to the EM update compared to the NR update.

Details

This is the Rcpp-coded Direct-Inversion LMM REML algorithm used behind the mmer function.

Value

If all parameters are correctly indicated the program will return a list with the following information:

Vi

the inverse of the phenotypic variance matrix V^- = (ZGZ+R)^-1

P

the projection matrix Vi - [Vi*(X*Vi*X)^-*Vi]

sigma

a list with the values of the variance-covariance components with one list element for each random effect.

sigma_scaled

a list with the values of the scaled variance-covariance components with one list element for each random effect.

sigmaSE

Hessian matrix containing the variance-covariance for the variance components. SE's can be obtained taking the square root of the diagonal values of the Hessian.

Beta

a data frame for trait BLUEs (fixed effects).

VarBeta

a variance-covariance matrix for trait BLUEs

U

a list (one element for each random effect) with a data frame for trait BLUPs.

VarU

a list (one element for each random effect) with the variance-covariance matrix for trait BLUPs.

PevU

a list (one element for each random effect) with the predicted error variance matrix for trait BLUPs.

fitted

Fitted values y.hat=XB

residuals

Residual values e = Y - XB

AIC

Akaike information criterion

BIC

Bayesian information criterion

convergence

a TRUE/FALSE statement indicating if the model converged.

monitor

The values of log-likelihood and variance-covariance components across iterations during the REML estimation.

percChange

The percent change of variance components across iterations. There should be one column less than the number of iterations. Calculated as percChange = ((x_i/x_i-1) - 1) * 100 where i is the ith iteration.

dL

The vector of first derivatives of the likelihood with respect to the ith variance-covariance component.

dL2

The matrix of second derivatives of the likelihood with respect to the i.j th variance-covariance component.

References

Mrode, R. A. (2014). Linear models for the prediction of animal breeding values. Cabi.

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

See also

Examples

data(DT_cpdata, package="enhancer")
DT <- DT_cpdata

# \donttest{
# response matrix
y <- cbind(imputev(DT$Yield), imputev(DT$Firmness))
# fixed effect incidence matrix
X <- model.matrix(~Rowf, data=DT)
# random effect incidence matrix
Z <- Matrix::sparse.model.matrix(~id - 1, data=DT)
colnames(Z) <- gsub("id","",colnames(Z))
# covariance for random effect
GT <- GT_cpdata
A <- A.mat(GT) # additive relationship matrix
A <- A[colnames(Z),colnames(Z)] # make sure of the order
A <- A + diag(1e-4,nrow(A), nrow(A))
# residual effect incidence matrix (dimensions equal nrow(y))
R1 <- Matrix::Diagonal(n=nrow(y)) 
# weights matrix (dimensions equal nrow(y))
W <- diag(nrow(y)) 
# some other parameters
maxIter=3
stepWeight <- rep(0.9, maxIter)
stepWeight[1:2] <- c(0.5, 0.7)
emWeights <- rep(0,maxIter)
# model fit
res <- MNR( Y=y, # multi-trait response
            X=list(X), Gx=list(diag(2)), # fixed effects
            Z=list(Z), K=list(A), # random effects
            R=list(R1), # residual effects
            Ge=list( (diag(2)*.3)+.15 , diag(2)*0.75 ), # inital vc 
            GeI=list(unsm2(2),diag(2)), # vc constraints
            W=W, isInvW=TRUE, # weights for records
            iters=maxIter,
            tolpar=1e-4, tolparinv=1e-6, # tolerance
            ai=FALSE, pev=FALSE, # algorithm specifics
            verbose=TRUE, 
            retscaled=FALSE, 
            stepweight=stepWeight, # second derivatives weights
            emweight=emWeights # em update weights
)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -278.58   21:17:37      1           0
#>     2      -273.609   21:17:38      2           0
#>     3      -271.388   21:17:39      3           0
res$sigma
#> , , 1
#> 
#>           [,1]      [,2]
#> [1,]  763.1123 -154.7015
#> [2,] -154.7015 1210.9463
#> 
#> , , 2
#> 
#>          [,1]     [,2]
#> [1,] 3210.848    0.000
#> [2,]    0.000 1103.532
#> 
res$Beta
#>            [,1]
#>  [1,]  77.71146
#>  [2,] 604.67646
#>  [3,]  20.04498
#>  [4,]  17.48948
#>  [5,]  53.52505
#>  [6,]  27.31167
#>  [7,]  56.38248
#>  [8,]  47.30879
#>  [9,]  36.91362
#> [10,]  40.57282
#> [11,]  60.45684
#> [12,]  43.64152
#> [13,] 117.34349
#> [14,]  35.68388
#> [15,]  80.19965
#> [16,]  45.10649
#> [17,]  58.07509
#> [18,]  49.11181
#> [19,]  95.95388
#> [20,]  30.76722
#> [21,]  48.09290
#> [22,]  30.84908
#> [23,]  31.13765
#> [24,]  34.33110
#> [25,]  49.79506
#> [26,]  42.21580
res$U[[1]]
#>                [,1]
#>   [1,] -15.07958599
#>   [2,]  -0.11647473
#>   [3,]   4.21680057
#>   [4,]   1.35313020
#>   [5,] -13.05837800
#>   [6,]  19.32986025
#>   [7,]  -4.26680025
#>   [8,]  -7.58944090
#>   [9,] -24.28811552
#>  [10,] -26.06152205
#>  [11,]  -6.83454855
#>  [12,]  30.73901409
#>  [13,]  33.96841343
#>  [14,]  13.02402712
#>  [15,]  20.83938298
#>  [16,]  13.60659507
#>  [17,]   7.29255492
#>  [18,]  -2.00581490
#>  [19,] -16.56678502
#>  [20,]  11.00987371
#>  [21,]  -4.43692871
#>  [22,] -66.73836180
#>  [23,]  21.19499292
#>  [24,]  10.21555286
#>  [25,]  12.70743448
#>  [26,]  10.09997790
#>  [27,]  14.08182278
#>  [28,]   9.88174567
#>  [29,]  21.05178977
#>  [30,] -33.30844881
#>  [31,]  16.06915235
#>  [32,] -17.87466848
#>  [33,]  21.84949635
#>  [34,] -14.05590143
#>  [35,]   6.04262674
#>  [36,]  16.72025924
#>  [37,]   2.88264783
#>  [38,]  11.03908190
#>  [39,]   0.52034144
#>  [40,]   0.93781001
#>  [41,]   1.31689201
#>  [42,] -46.84490124
#>  [43,] -16.45900406
#>  [44,]  59.49725282
#>  [45,] -12.22586628
#>  [46,]  21.19795065
#>  [47,]  12.69939377
#>  [48,]   2.21012039
#>  [49,]   3.72867401
#>  [50,] -25.61819783
#>  [51,]  13.11178305
#>  [52,] -16.29663115
#>  [53,] -18.29504208
#>  [54,] -23.18601683
#>  [55,] -28.23768209
#>  [56,]  49.54228673
#>  [57,]  -7.91005549
#>  [58,] -38.89263900
#>  [59,]  15.40348297
#>  [60,]  10.69757065
#>  [61,]  12.49303945
#>  [62,] -25.04693291
#>  [63,]   6.47786419
#>  [64,]  18.84092813
#>  [65,]  18.09833889
#>  [66,] -15.29857564
#>  [67,] -11.68458362
#>  [68,]  21.26275710
#>  [69,]  15.33769920
#>  [70,] -30.18993950
#>  [71,]  10.40944275
#>  [72,]   3.51835900
#>  [73,]   9.43140305
#>  [74,]  12.35150328
#>  [75,]   3.62076537
#>  [76,] -22.48395276
#>  [77,]  15.70666243
#>  [78,] -45.67900238
#>  [79,]  27.55800047
#>  [80,]  -6.81964907
#>  [81,]   1.23122015
#>  [82,]  12.09331742
#>  [83,]  41.71456062
#>  [84,]  -1.71230159
#>  [85,] -14.78474164
#>  [86,]  38.96419385
#>  [87,]  19.80259658
#>  [88,]   4.99000678
#>  [89,]  -9.40473288
#>  [90,]  -0.78750127
#>  [91,]   7.27753750
#>  [92,] -39.40525527
#>  [93,]  -5.96022601
#>  [94,]  -6.53606250
#>  [95,]  25.20349463
#>  [96,] -18.92802456
#>  [97,]  -1.67239082
#>  [98,]  13.56573569
#>  [99,]  -2.53545786
#> [100,]  44.13296833
#> [101,] -19.97945101
#> [102,] -27.35580019
#> [103,] -10.65710041
#> [104,]  -4.24075015
#> [105,]   5.29850301
#> [106,] -24.33866046
#> [107,]  20.39301275
#> [108,] -22.62916004
#> [109,]   6.39751682
#> [110,]  34.68500934
#> [111,]   5.77097567
#> [112,] -11.03745235
#> [113,]  -0.99504885
#> [114,] -23.49847156
#> [115,]  12.02080213
#> [116,] -21.34142779
#> [117,]  19.33982679
#> [118,] -14.94045049
#> [119,]  14.28072341
#> [120,]  52.73691182
#> [121,]  -0.62549956
#> [122,]  -1.20728332
#> [123,] -11.41904583
#> [124,]  -3.99495977
#> [125,]  -0.81915122
#> [126,]  13.93449934
#> [127,] -21.67517430
#> [128,]  56.12152438
#> [129,]   0.19630773
#> [130,]  13.34882308
#> [131,]  -7.32167022
#> [132,]  17.48065868
#> [133,]   8.85685612
#> [134,] -29.02797135
#> [135,]   8.58085316
#> [136,]   5.64418456
#> [137,]  -9.65265496
#> [138,]  19.63221059
#> [139,] -12.72744485
#> [140,]  22.86447886
#> [141,]  10.37731104
#> [142,]  34.55348095
#> [143,]   5.26121264
#> [144,]  13.71205275
#> [145,]  -1.89475309
#> [146,]  34.55506015
#> [147,]  24.19724544
#> [148,] -47.68080755
#> [149,] -10.64202367
#> [150,]  13.27725104
#> [151,]   6.16679770
#> [152,]   5.92800106
#> [153,]   0.04362395
#> [154,]   5.64359545
#> [155,]  -4.76382459
#> [156,] -21.47542850
#> [157,]  -5.80067729
#> [158,] -33.42959224
#> [159,]  10.72636559
#> [160,] -15.26170504
#> [161,]   7.42540289
#> [162,]  35.40709302
#> [163,]  11.42119276
#> [164,]  27.68436951
#> [165,]   8.15954535
#> [166,]  -1.59480825
#> [167,]  -8.27377331
#> [168,]  -2.63891045
#> [169,]  37.86850015
#> [170,] -17.79601861
#> [171,]   0.60299328
#> [172,]  -7.66814621
#> [173,]  10.53275078
#> [174,] -10.16488480
#> [175,]  -9.16912691
#> [176,]  -1.64463356
#> [177,] -23.25039527
#> [178,]  -8.63825645
#> [179,] -13.75925516
#> [180,]  12.29693239
#> [181,]  -9.07687860
#> [182,] -38.48969985
#> [183,]  18.56648246
#> [184,]  -3.98727060
#> [185,] -24.70242354
#> [186,] -19.66105056
#> [187,]   5.03938684
#> [188,]  22.62086995
#> [189,]   0.42319145
#> [190,]  35.47399234
#> [191,] -28.61288062
#> [192,]  -6.00705911
#> [193,]  15.34882172
#> [194,] -15.54693278
#> [195,]  18.97995131
#> [196,] -14.28355658
#> [197,]  -9.90270833
#> [198,] -25.80832943
#> [199,]  25.02001278
#> [200,] -53.87853317
#> [201,]   6.75946574
#> [202,]  -4.63087183
#> [203,]  -8.21578391
#> [204,]  -7.71054422
#> [205,] -13.27095537
#> [206,]   4.28793511
#> [207,]   2.79068402
#> [208,]  12.22563870
#> [209,]   8.65729675
#> [210,]  -3.23129560
#> [211,]  14.92794198
#> [212,] -10.45594341
#> [213,]  -5.54358965
#> [214,]  12.29479449
#> [215,]  17.88632552
#> [216,]   6.98280106
#> [217,]  -2.44263459
#> [218,] -20.00451134
#> [219,]  -7.16745624
#> [220,]   9.51208480
#> [221,] -18.17379062
#> [222,]   9.68841583
#> [223,]  -5.20584477
#> [224,] -16.03339238
#> [225,] -33.05194119
#> [226,] -15.93109782
#> [227,]   6.59849721
#> [228,]  41.20869389
#> [229,]  22.29347259
#> [230,] -21.02169432
#> [231,] -20.20682271
#> [232,] -32.05997301
#> [233,]  -6.47918581
#> [234,] -29.87663323
#> [235,]  -4.50638813
#> [236,]  12.15549568
#> [237,] -16.64763722
#> [238,]  25.34932366
#> [239,]  -5.44300643
#> [240,]  -7.95256307
#> [241,]  -7.94951472
#> [242,]  -9.79080383
#> [243,] -36.42815616
#> [244,]  31.78275171
#> [245,]  -1.82661074
#> [246,]  -5.24468707
#> [247,]  12.35449179
#> [248,] -33.06837611
#> [249,]  39.68996619
#> [250,] -39.25805637
#> [251,]  40.51204145
#> [252,] -32.13376428
#> [253,] -22.60538692
#> [254,]  -2.02140211
#> [255,]  12.28152646
#> [256,] -12.62599502
#> [257,]   4.54687728
#> [258,]  -2.57690032
#> [259,]   2.31334877
#> [260,] -28.51059631
#> [261,]  -1.47089586
#> [262,]  33.05571579
#> [263,]   6.24252545
#> [264,] -18.44520722
#> [265,]  -5.99917162
#> [266,] -14.35706894
#> [267,]  -8.09945365
#> [268,] -11.68216328
#> [269,]  -0.48234164
#> [270,]  -2.09642356
#> [271,] -23.69037533
#> [272,]  -8.44111699
#> [273,]  -7.68589734
#> [274,]   3.72257827
#> [275,] -24.33467431
#> [276,]   2.24920379
#> [277,]  -0.48327056
#> [278,]  -2.37175118
#> [279,] -20.67253707
#> [280,]   6.23959563
#> [281,]  -6.94604598
#> [282,]  78.53711243
#> [283,]   2.83890237
#> [284,] -32.04737749
#> [285,] -17.58187698
#> [286,]   9.55479928
#> [287,]   1.62425025
#> [288,]  -5.64312986
#> [289,]  -0.47921555
#> [290,]  29.71071689
#> [291,]  -7.84715578
#> [292,]  10.79138851
#> [293,]  17.14657921
#> [294,]  15.45510859
#> [295,]  35.92430097
#> [296,] -42.21808457
#> [297,]  17.38101638
#> [298,]  38.67088962
#> [299,]   6.57468977
#> [300,]  19.02277327
#> [301,] -19.71212270
#> [302,] -12.16777969
#> [303,] -18.36839842
#> [304,]  -8.26758350
#> [305,] -27.19100756
#> [306,] -16.74135283
#> [307,]  11.55292358
#> [308,]   7.34230470
#> [309,] -20.24424082
#> [310,] -30.34140508
#> [311,] -32.53798543
#> [312,] -23.52190643
#> [313,]  13.98709737
#> [314,] -16.68567347
#> [315,]  -2.71225622
#> [316,]  24.32296109
#> [317,]  -3.22116801
#> [318,]  -4.08944398
#> [319,]   2.74334525
#> [320,]  10.87400920
#> [321,]  -1.15316859
#> [322,]  -2.75735023
#> [323,]  29.07965717
#> [324,] -38.70543432
#> [325,]  12.26749522
#> [326,]   2.31157482
#> [327,] -31.46107250
#> [328,]  -3.51085488
#> [329,] -24.41570848
#> [330,] -19.70462610
#> [331,]   6.41585447
#> [332,] -13.66657267
#> [333,]   7.44821189
#> [334,]  18.50464670
#> [335,]  18.67836482
#> [336,]  26.35644969
#> [337,]   0.86642536
#> [338,] -19.17966735
#> [339,] -26.52465309
#> [340,]  26.62771630
#> [341,]   5.82555365
#> [342,] -30.23625395
#> [343,]  38.35314063
#> [344,] -16.08270245
#> [345,]  -3.59310110
#> [346,]  -4.97275022
#> [347,] -14.27445445
#> [348,] -27.79702321
#> [349,] -25.10362949
#> [350,]  42.65571640
#> [351,]  -8.04896972
#> [352,]  13.34968046
#> [353,] -16.29645927
#> [354,]   9.74604884
#> [355,]  32.07384189
#> [356,] -23.14343910
#> [357,]  -3.99452368
#> [358,] -32.81939215
#> [359,] -18.61039321
#> [360,]  49.08973585
#> [361,]   6.56028850
#> [362,]  10.48361550
#> [363,] -24.17806131
#> [364,]  -8.04976311
#> [365,]  -2.13145270
#> [366,]  12.05743010
#> [367,] -33.45433939
#> [368,]   0.34945470
#> [369,]   7.58934245
#> [370,] -26.45343769
#> [371,]   9.21972768
#> [372,] -11.71371616
#> [373,] -12.16121196
#> [374,] -33.34574992
#> [375,]  18.86903327
#> [376,] -10.60995684
#> [377,] -10.62577466
#> [378,] -18.70334395
#> [379,]  10.36172024
#> [380,] -13.87972475
#> [381,] -19.58506480
#> [382,]  60.23582424
#> [383,] -28.46169779
#> [384,]   5.38123661
#> [385,]  17.33784592
#> [386,]  25.76799581
#> [387,]   2.19664108
#> [388,] -12.98168662
#> [389,] -10.26301430
#> [390,]  -4.48341079
#> [391,]   2.74777429
#> [392,]  25.04900568
#> [393,]  -7.47060330
#> [394,]  40.39319696
#> [395,]  -2.64934783
#> [396,]   7.22018002
#> [397,]  13.35550572
#> [398,]  15.13778305
#> [399,]  13.02940719
#> [400,]  34.44060916
#> [401,]  29.16062271
#> [402,]   6.19179673
#> [403,]   1.98124975
#> [404,]  18.00785786
#> [405,]  18.14000684
#> [406,] -23.43130029
#> [407,] -30.93763844
#> [408,]  43.01364623
#> [409,]  -4.25016111
#> [410,] -12.49971832
#> [411,] -23.11834001
#> [412,]   3.02215710
#> [413,]  20.94999320
#> [414,] -13.27806985
#> [415,]  10.26047671
#> [416,] -26.62299243
#> [417,]  -1.24520558
#> [418,] -29.98487118
#> [419,]   1.87469809
#> [420,]  15.10329526
#> [421,]  -7.39943759
#> [422,] -46.13344855
#> [423,] -12.59103297
#> [424,]  36.41588819
#> [425,] -17.08423613
#> [426,]  28.47316528
#> [427,]  24.62096716
#> [428,] -16.90517689
#> [429,]   4.33361928
#> [430,]   5.25551618
#> [431,] -44.04290263
#> [432,]  45.47261553
#> [433,]  -4.73363750
#> [434,]  35.91136833
#> [435,]  17.72993140
#> [436,] -28.20883757
#> [437,]   5.62809438
#> [438,]   2.35936103
#> [439,]  14.51140123
#> [440,]   6.28228111
#> [441,]  -4.11161159
#> [442,]  24.01020986
#> [443,]   3.65059606
#> [444,] -30.84680860
#> [445,]  13.14199789
#> [446,] -20.46181126
#> [447,]  24.85822212
#> [448,] -16.93895775
#> [449,] -15.27030811
#> [450,] -22.77613478
#> [451,] -35.85066512
#> [452,]  20.04005091
#> [453,]   9.76998425
#> [454,] -42.54821984
#> [455,]  -8.91178022
#> [456,] -31.57503312
#> [457,]  35.02018546
#> [458,]  26.60382480
#> [459,]   4.94369853
#> [460,] -32.07249269
#> [461,]   5.34391632
#> [462,]  41.37249333
#> [463,]  21.79611448
#> [464,]  -8.87511230
#> [465,]   3.53477969
#> [466,]  -4.50762284
#> [467,]  17.75120956
#> [468,]  29.55076888
#> [469,]   1.07854398
#> [470,]  32.77403779
#> [471,] -32.57047285
#> [472,]  17.07165805
#> [473,]   6.17140642
#> [474,] -52.53944404
#> [475,]   1.70730662
#> [476,]  -4.13000175
#> [477,]  -9.95948225
#> [478,] -13.81899657
#> [479,]  -0.94967051
#> [480,]   2.99540642
#> [481,]  -4.87074003
#> [482,] -14.84734869
#> [483,]  -2.59255331
#> [484,]  34.69315174
#> [485,]   3.22560695
#> [486,]  -7.39040698
#> [487,]   5.20520460
#> [488,]  17.43054370
#> [489,]   0.70342424
#> [490,]  14.55296838
#> [491,] -21.14777342
#> [492,]  -9.19030662
#> [493,]  29.39400927
#> [494,] -30.45420128
#> [495,]   8.04265554
#> [496,]  17.45608641
#> [497,]  10.68582638
#> [498,]  28.46649424
#> [499,] -21.75599957
#> [500,] -19.54432777
#> [501,]  21.51091663
#> [502,]  20.88007405
#> [503,]  -9.76667958
#> [504,] -78.23422055
#> [505,]   6.03862809
#> [506,]  19.77210263
#> [507,]   5.99090221
#> [508,]  14.07294365
#> [509,]  13.75961734
#> [510,]   9.56622395
#> [511,]  14.64068357
#> [512,] -32.22951151
#> [513,] -35.92340009
#> [514,]  34.34430472
#> [515,]  -6.05910203
#> [516,]  23.59141052
#> [517,]  -6.99189179
#> [518,]   8.60560814
#> [519,]   9.97176422
#> [520,]  27.69699159
#> [521,]  21.12022568
#> [522,] -12.49736446
#> [523,]   3.14305810
#> [524,]  27.46797224
#> [525,] -15.44474047
#> [526,]  31.03941731
#> [527,]  32.29713713
#> [528,]   2.96320713
#> [529,]  -1.84996447
#> [530,]  -8.24716596
#> [531,]   0.35749804
#> [532,]  41.96912542
#> [533,]   7.16457582
#> [534,]  32.28665685
#> [535,]   5.16686423
#> [536,]   6.03569358
#> [537,]  -8.03004507
#> [538,]   4.09325894
#> [539,]   3.61629732
#> [540,]   8.60854542
#> [541,]  -8.40558508
#> [542,] -17.54444898
#> [543,] -15.92824408
#> [544,] -21.50465783
#> [545,]  11.76810334
#> [546,] -22.30148789
#> [547,] -29.08419894
#> [548,]  -2.87933698
#> [549,] -12.65444265
#> [550,] -40.10698049
#> [551,] -14.52204689
#> [552,]   6.84727667
#> [553,]  26.00998789
#> [554,] -18.46690448
#> [555,] -15.82197002
#> [556,] -20.02845041
#> [557,] -25.78188681
#> [558,]  24.05209512
#> [559,]  -5.56746674
#> [560,]   3.26544598
#> [561,] -38.27665864
#> [562,]  13.86114821
#> [563,]  19.94979006
#> [564,]  -2.65061386
#> [565,]   6.40495659
#> [566,]  -3.02466289
#> [567,]   5.54507836
#> [568,]  13.54662708
#> [569,]  -0.90337446
#> [570,] -21.90920488
#> [571,]  -4.97145322
#> [572,]   2.43772948
#> [573,] -13.47231440
#> [574,]  -5.98789678
#> [575,]  -2.88683022
#> [576,]  14.98892756
#> [577,]  26.52199047
#> [578,]   4.78616429
#> [579,]   2.42276142
#> [580,] -20.61762491
#> [581,] -22.20403654
#> [582,]  54.84560143
#> [583,]  -4.92330299
#> [584,] -25.52054403
#> [585,]   5.04685953
#> [586,] -11.30816493
#> [587,]  -7.30529401
#> [588,]  25.74408318
#> [589,]  10.20103763
#> [590,]  -8.56179528
#> [591,] -10.40309197
#> [592,]  11.31143786
#> [593,]   9.91972047
#> [594,]  17.32915026
#> [595,]  -3.00920511
#> [596,] -48.27525551
#> [597,] -17.94433185
#> [598,] -48.74090061
#> [599,]  -0.32524004
#> [600,]  46.27518021
#> [601,] -13.02732652
#> [602,] -13.98445304
#> [603,]  -0.86493570
#> [604,]  18.52907571
#> [605,]  -6.59130744
#> [606,] -20.33256908
#> [607,]   3.80980327
#> [608,]  46.83911364
#> [609,]   2.55817473
#> [610,]  -9.60657203
#> [611,]  12.37872917
#> [612,]   0.57322098
#> [613,]  -4.34294969
#> [614,]  16.78142482
#> [615,]   7.74593394
#> [616,] -21.39461491
#> [617,]  13.63410886
#> [618,]  13.77163097
#> [619,]  36.90623517
#> [620,] -20.67535305
#> [621,]  12.25886225
#> [622,] -29.01226042
#> [623,]   3.99510267
#> [624,]  40.36725008
#> [625,]  29.15933588
#> [626,]   0.09860014
#> [627,]  13.09318747
#> [628,]  -7.26828802
#> [629,] -11.14777546
#> [630,]   0.81536857
#> [631,]  14.04154918
#> [632,]  14.25007638
#> [633,]   7.60062018
#> [634,] -10.36477623
#> [635,]  33.12949080
#> [636,]  21.02252423
#> [637,] -20.66616394
#> [638,]  -1.70243550
#> [639,] -31.92776023
#> [640,]  15.83179073
#> [641,]  23.76931989
#> [642,] -45.54105130
#> [643,]   8.53126517
#> [644,] -15.22077363
#> [645,]  18.32157244
#> [646,]  20.77080145
#> [647,]   1.95414936
#> [648,]  15.23718919
#> [649,]  30.26122602
#> [650,] -23.52855818
#> [651,]   2.36582464
#> [652,]  12.78650271
#> [653,] -16.08015547
#> [654,]  -1.84665561
#> [655,]  35.10022317
#> [656,]  -4.27868891
#> [657,]   8.12708493
#> [658,]   0.61150570
#> [659,] -19.51683271
#> [660,]   8.87096298
#> [661,]  -9.18732624
#> [662,]  15.81536622
#> [663,] -24.75227837
#> [664,]  -5.26921551
#> [665,]  -3.43982911
#> [666,]   1.16395273
#> [667,]   1.09263629
#> [668,]  -5.33369254
#> [669,]  -5.56312535
#> [670,]  25.74001813
#> [671,]  -8.13298492
#> [672,] -15.22130841
#> [673,] -17.75144910
#> [674,]   4.94325413
#> [675,]  -1.16342653
#> [676,] -22.03225293
#> [677,]   5.84763077
#> [678,]  -4.59194138
#> [679,]  -3.70951247
#> [680,] -21.26763202
#> [681,] -18.02603701
#> [682,] -12.79481985
#> [683,] -32.61151894
#> [684,]   2.50445682
#> [685,]  -8.56290657
#> [686,]  13.89380661
#> [687,]  37.55435694
#> [688,]  52.06644436
#> [689,]  -9.60458083
#> [690,]  15.13151153
#> [691,] -22.88107138
#> [692,]  -0.97710786
#> [693,]   3.64076222
#> [694,]  24.46064335
#> [695,]  -3.71788511
#> [696,]  18.57288670
#> [697,] -12.87211706
#> [698,]   7.09170820
#> [699,] -10.67339451
#> [700,] -24.16725486
#> [701,]  -4.89008009
#> [702,]  -9.41335406
#> [703,]   2.90276873
#> [704,]  -0.64613093
#> [705,] -20.97487200
#> [706,]  22.57351125
#> [707,] -42.35320630
#> [708,]  11.01596233
#> [709,]   3.65813556
#> [710,]  -3.68288857
#> [711,] -29.24519673
#> [712,]   4.44714885
#> [713,]   4.76989356
#> [714,]   0.66144472
#> [715,]  -5.07184699
#> [716,]  20.42606284
#> [717,]   8.34750750
#> [718,]  -3.78922620
#> [719,] -19.38485364
#> [720,] -19.37142928
#> [721,]  12.09042591
#> [722,] -23.21407067
#> [723,] -25.21867314
#> [724,]  29.34396719
#> [725,]  11.63721963
#> [726,]   3.10776215
# }