covariance between random effects

covc merges the incidence matrices and covariance matrices of two random effects to fit an unstructured model between 2 different random effects to be fitted with the mmec solver.

Usage

covc(ran1, ran2, thetaC=NULL, theta=NULL)

Arguments

ran1

the random call of the first random effect.

ran2

the random call of the first random effect.

thetaC

an optional matrix for constraints in the variance components.

theta

an optional symmetric matrix for initial values of the variance-covariance components. When providing customized values, these values should be scaled with respect to the original variance. For example, to provide an initial value of 1 to a given variance component, theta would be built as:

theta = matrix( 1 / var(response) )

The symmetric matrix should have as many rows and columns as the number of levels in the factor 'x'. The values in the matrix define the initial values of the variance-covariance components that will be subject to the constraints provided in thetaC. If not provided, initial values will be calculated as:

theta = diag(ncol(mm))*.05 + matrix(.1,ncol(mm),ncol(mm))

where mm is the incidence matrix for the factor 'x'.

Details

This implementation aims to fit models where covariance between random variables is expected to exist. For example, indirect genetic effects.

Value

$Z

a incidence matrix Z* = Z Gamma which is the original incidence matrix for the timevar multiplied by the loadings.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Bijma, P. (2014). The quantitative genetics of indirect genetic effects: a selective review of modelling issues. Heredity, 112(1), 61-69.

Author

Giovanny Covarrubias-Pazaran

See also

Examples

data(DT_ige)
DT <- DT_ige
covRes <- with(DT, covc( vsc(isc(focal)) , vsc(isc(neighbour)) ) )
str(covRes)
#> List of 7
#>  $ Z          :List of 2
#>   ..$ :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. .. ..@ i       : int [1:1000] 305 313 542 567 605 619 626 717 953 131 ...
#>   .. .. ..@ p       : int [1:99] 0 9 14 20 22 25 28 32 86 91 ...
#>   .. .. ..@ Dim     : int [1:2] 1000 98
#>   .. .. ..@ Dimnames:List of 2
#>   .. .. .. ..$ : chr [1:1000] "1" "2" "3" "4" ...
#>   .. .. .. ..$ : chr [1:98] "id_1019" "id_1101" "id_1137" "id_1268" ...
#>   .. .. ..@ x       : num [1:1000] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..@ factors : list()
#>   ..$ :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. .. ..@ i       : int [1:1000] 350 567 591 593 595 717 755 859 105 200 ...
#>   .. .. ..@ p       : int [1:99] 0 8 14 21 24 28 30 32 84 88 ...
#>   .. .. ..@ Dim     : int [1:2] 1000 98
#>   .. .. ..@ Dimnames:List of 2
#>   .. .. .. ..$ : chr [1:1000] "1" "2" "3" "4" ...
#>   .. .. .. ..$ : chr [1:98] "id_1019" "id_1101" "id_1137" "id_1268" ...
#>   .. .. ..@ x       : num [1:1000] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..@ factors : list()
#>  $ Gu         :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:98] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ p       : int [1:99] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ Dim     : int [1:2] 98 98
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:98] "id_1019" "id_1101" "id_1137" "id_1268" ...
#>   .. .. ..$ : chr [1:98] "id_1019" "id_1101" "id_1137" "id_1268" ...
#>   .. ..@ x       : num [1:98] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..@ factors : list()
#>  $ theta      : num [1:2, 1:2] 0.165 0.015 0.015 0.165
#>  $ thetaC     : num [1:2, 1:2] 1 0 2 1
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "ran1" "ran2"
#>   .. ..$ : chr [1:2] "ran1" "ran2"
#>  $ thetaF     : num[0 , 0 ] 
#>  $ partitionsR: list()
#>  $ sp         : num(0) 
# look at DT_ige help page to see how to fit an actual model