Imputing a numeric or character vector

This function is a very simple function to impute a numeric or character vector with the mean or median value of the vector.

Usage

imputev(x, method="median")

Arguments

x

a numeric or character vector.

method

the method to choose between mean or median.

Value

$x

a numeric or character vector imputed with the method selected.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Author

Giovanny Covarrubias-Pazaran

See also

Examples

####=========================================####
#### generate your mickey mouse -log10(p-values)
####=========================================####
set.seed(1253)
x <- rnorm(100)
x[sample(1:100,10)] <- NA
imputev(x)
#>   [1]  1.577855068  0.240143762 -1.331277824  1.025186918 -0.928600204
#>   [6]  0.694531876 -0.744847948  1.655195579  0.044981432  0.784703034
#>  [11] -1.020295637  1.174174050  0.044981432  0.457794841  1.257489611
#>  [16] -0.412769854  0.635808774  0.545925718  0.163391748  0.297052560
#>  [21]  0.564450938 -0.316481020 -0.555029207  1.004811343  0.044981432
#>  [26] -1.924113615 -0.927344504  0.995348447  1.048837095 -0.155236834
#>  [31]  0.950410434  1.160607126  0.044981432  0.195792188  1.112691701
#>  [36]  1.204521218 -2.910366676  0.748453777 -0.694321858  0.044981432
#>  [41]  1.685779213 -0.074461704  0.152088508  0.540014238 -0.677144037
#>  [46]  0.209193559 -0.055097378  2.218282071 -1.991467171 -2.367082599
#>  [51] -0.520100402 -0.188628144 -1.212232668 -0.051738747  1.359198689
#>  [56]  0.044981432  1.028300535  0.038970031 -0.053774901 -0.287839774
#>  [61]  0.044981432 -0.320252863  0.715899389 -1.804831676  2.133901448
#>  [66] -0.219905932  0.054745279 -0.369139043 -0.437421760 -0.493611343
#>  [71] -0.641825613 -1.033010954 -0.561891136  1.472184031  0.356134347
#>  [76]  0.050992833 -0.575716084 -0.836216728  1.455376860  0.714424589
#>  [81]  0.705273802  0.044981432 -1.787915833 -0.278508709  0.103233527
#>  [86] -1.255143163  1.832895750 -0.367893289 -1.560032985 -0.752854983
#>  [91]  0.003281394  0.044981432 -1.147504621  0.186033348  0.044981432
#>  [96]  0.526298279  0.428082812  0.175179502 -0.211427533 -0.129934458