list or vector to unstructured matrix

list2usmat creates an unstructured square matrix taking a vector or list to fill the diagonal and upper triangular with the values provided.

Usage

list2usmat(sigmaL)

Arguments

sigmaL

vector or list of values to put on the matrix.

Value

$res

a matrix with the values provided.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Author

Giovanny Covarrubias-Pazaran

See also

Examples

list2usmat(as.list(1:3))
#>      [,1] [,2]
#> [1,]    1    2
#> [2,]    2    3
list2usmat(as.list(1:10))
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    2    4    7
#> [2,]    2    3    5    8
#> [3,]    4    7    6    9
#> [4,]    5    8    9   10