Overlay Matrix

`overlay` adds r times the design matrix for model term t to the existing design matrix. Specifically, if the model up to this point has p effects and t has a effects, the a columns of the design matrix for t are multiplied by the scalar r (default value 1.0). This can be used to force a correlation of 1 between two terms as in a diallel analysis.

Usage

overlay(..., rlist=NULL, prefix=NULL, sparse=FALSE)

Arguments

...

as many vectors as desired to overlay.

rlist

a list of scalar values indicating the times that each incidence matrix overlayed should be multiplied by. By default r=1.

prefix

a character name to be added before the column names of the final overlay matrix. This may be useful if you have entries with names starting with numbers which programs such as asreml doesn't like, or for posterior extraction of parameters, that way 'grep'ing is easier.

sparse

a TRUE/FALSE statement specifying if the matrices should be built as sparse or regular matrices.

Value

$S3

an incidence matrix with as many columns levels in the vectors provided to build the incidence matrix.

References

Fikret Isik. 2009. Analysis of Diallel Mating Designs. North Carolina State University, Raleigh, USA.

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Author

Giovanny Covarrubias-Pazaran

See also

Examples

  
####=========================================####
#### For CRAN time limitations most lines in the 
#### examples are silenced with one '#' mark, 
#### remove them and run the examples
####=========================================####
data("DT_halfdiallel")
DT <- DT_halfdiallel
head(DT)
#>   rep geno male female     sugar
#> 1   1   12    1      2 13.950509
#> 2   2   12    1      2  9.756918
#> 3   1   13    1      3 13.906355
#> 4   2   13    1      3  9.119455
#> 5   1   14    1      4  5.174483
#> 6   2   14    1      4  8.452221
DT$femalef <- as.factor(DT$female)
DT$malef <- as.factor(DT$male)
DT$genof <- as.factor(DT$geno)

with(DT, overlay(femalef,malef, sparse = TRUE))
#> 42 x 7 sparse Matrix of class "dgCMatrix"
#>    1 2 3 4 5 6 7
#> 1  1 1 . . . . .
#> 2  1 1 . . . . .
#> 3  1 . 1 . . . .
#> 4  1 . 1 . . . .
#> 5  1 . . 1 . . .
#> 6  1 . . 1 . . .
#> 7  1 . . . 1 . .
#> 8  1 . . . 1 . .
#> 9  1 . . . . 1 .
#> 10 1 . . . . 1 .
#> 11 1 . . . . . 1
#> 12 1 . . . . . 1
#> 13 . 1 1 . . . .
#> 14 . 1 1 . . . .
#> 15 . 1 . 1 . . .
#> 16 . 1 . 1 . . .
#> 17 . 1 . . 1 . .
#> 18 . 1 . . 1 . .
#> 19 . 1 . . . 1 .
#> 20 . 1 . . . 1 .
#> 21 . 1 . . . . 1
#> 22 . 1 . . . . 1
#> 23 . . 1 1 . . .
#> 24 . . 1 1 . . .
#> 25 . . 1 . 1 . .
#> 26 . . 1 . 1 . .
#> 27 . . 1 . . 1 .
#> 28 . . 1 . . 1 .
#> 29 . . 1 . . . 1
#> 30 . . 1 . . . 1
#> 31 . . . 1 1 . .
#> 32 . . . 1 1 . .
#> 33 . . . 1 . 1 .
#> 34 . . . 1 . 1 .
#> 35 . . . 1 . . 1
#> 36 . . . 1 . . 1
#> 37 . . . . 1 1 .
#> 38 . . . . 1 1 .
#> 39 . . . . 1 . 1
#> 40 . . . . 1 . 1
#> 41 . . . . . 1 1
#> 42 . . . . . 1 1
with(DT, overlay(femalef,malef, sparse = FALSE))
#>    1 2 3 4 5 6 7
#> 1  1 1 0 0 0 0 0
#> 2  1 1 0 0 0 0 0
#> 3  1 0 1 0 0 0 0
#> 4  1 0 1 0 0 0 0
#> 5  1 0 0 1 0 0 0
#> 6  1 0 0 1 0 0 0
#> 7  1 0 0 0 1 0 0
#> 8  1 0 0 0 1 0 0
#> 9  1 0 0 0 0 1 0
#> 10 1 0 0 0 0 1 0
#> 11 1 0 0 0 0 0 1
#> 12 1 0 0 0 0 0 1
#> 13 0 1 1 0 0 0 0
#> 14 0 1 1 0 0 0 0
#> 15 0 1 0 1 0 0 0
#> 16 0 1 0 1 0 0 0
#> 17 0 1 0 0 1 0 0
#> 18 0 1 0 0 1 0 0
#> 19 0 1 0 0 0 1 0
#> 20 0 1 0 0 0 1 0
#> 21 0 1 0 0 0 0 1
#> 22 0 1 0 0 0 0 1
#> 23 0 0 1 1 0 0 0
#> 24 0 0 1 1 0 0 0
#> 25 0 0 1 0 1 0 0
#> 26 0 0 1 0 1 0 0
#> 27 0 0 1 0 0 1 0
#> 28 0 0 1 0 0 1 0
#> 29 0 0 1 0 0 0 1
#> 30 0 0 1 0 0 0 1
#> 31 0 0 0 1 1 0 0
#> 32 0 0 0 1 1 0 0
#> 33 0 0 0 1 0 1 0
#> 34 0 0 0 1 0 1 0
#> 35 0 0 0 1 0 0 1
#> 36 0 0 0 1 0 0 1
#> 37 0 0 0 0 1 1 0
#> 38 0 0 0 0 1 1 0
#> 39 0 0 0 0 1 0 1
#> 40 0 0 0 0 1 0 1
#> 41 0 0 0 0 0 1 1
#> 42 0 0 0 0 0 1 1
#> attr(,"variables")
#> [1] "femalef" "malef"