Predict form of a LMM fitted with mmec

predict method for class "mmec".

Usage

# S3 method for class 'mmec'
predict(object, Dtable=NULL, D, ...)

Arguments

object

a mixed model of class "mmec"

Dtable

a table specifying the terms to be included or averaged.

An "include" term means that the model matrices for that fixed or random effect is filled with 1's for the positions where column names and row names match.

An "include and average" term means that the model matrices for that fixed or random effect is filled with 1/#1's in that row.

An "average" term alone means that all rows for such fixed or random effect will be filled with 1/#levels in the effect.

If a term is not considered "include" or "average" is then totally ignored in the BLUP and SE calculation.

The default rule to invoke when the user doesn't provide the Dtable is to include and average all terms that match the argument D.

D

a character string specifying the variable used to extract levels for the rows of the D matrix and its construction. Alternatively, the D matrix (of class dgCMatrix) specifying the matrix to be used for the predictions directly.

...

Further arguments to be passed.

Details

This function allows to produce predictions specifying those variables that define the margins of the hypertable to be predicted (argument D). Predictions are obtained for each combination of values of the specified variables that is present in the data set used to fit the model. See vignettes for more details.

For predicted values the pertinent design matrices X and Z together with BLUEs (b) and BLUPs (u) are multiplied and added together.

predicted.value equal Xb + Zu.1 + ... + Zu.n

For computing standard errors for predictions the parts of the coefficient matrix:

C11 equal (X.t() V.inv() X).inv()

C12 equal 0 - [(X.t() V.inv() X).inv() X.t() V.inv() G Z]

C22 equal PEV equal G - [Z.t() G[V.inv() - (V.inv() X X.t() V.inv() X V.inv() X)]G Z.t()]

In practive C equals ( W.t() V.inv() W ).inv()

when both fixed and random effects are present in the inclusion set. If only fixed and random effects are included, only the respective terms from the SE for fixed or random effects are calculated.

Value

pvals

the table of predictions according to the specified arguments.

vcov

the variance covariance for the predictions.

D

the model matrix for predictions as defined in Welham et al.(2004).

Dtable

the table specifying the terms to include and terms to be averaged.

References

Welham, S., Cullis, B., Gogel, B., Gilmour, A., and Thompson, R. (2004). Prediction in linear mixed models. Australian and New Zealand Journal of Statistics, 46, 325 - 347.

Author

Giovanny Covarrubias-Pazaran

See also

Examples

data(DT_yatesoats)
DT <- DT_yatesoats
m3 <- mmec(fixed=Y ~ V + N + V:N ,
           random = ~ B + B:MP,
           rcov=~units,
           data = DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -220.267   1:16:20      0           0
#>     2      -213.15   1:16:20      0           0
#>     3      -209.956   1:16:20      0           0
#>     4      -209.412   1:16:20      0           0
#>     5      -209.384   1:16:20      0           0
#>     6      -209.379   1:16:20      0           0
#>     7      -209.378   1:16:20      0           0
#>     8      -209.378   1:16:20      0           0

#############################
## predict means for nitrogen
#############################
Dt <- m3$Dtable; Dt
#>     type               term include average
#> 1  fixed                  V   FALSE   FALSE
#> 2  fixed                  N   FALSE   FALSE
#> 3  fixed                V:N   FALSE   FALSE
#> 4 random    vsc( isc( B ) )   FALSE   FALSE
#> 5 random vsc( isc( B:MP ) )   FALSE   FALSE
# first fixed effect just average
Dt[1,"average"] = TRUE
# second fixed effect include
Dt[2,"include"] = TRUE
# third fixed effect include and average
Dt[3,"include"] = TRUE
Dt[3,"average"] = TRUE
Dt
#>     type               term include average
#> 1  fixed                  V   FALSE    TRUE
#> 2  fixed                  N    TRUE   FALSE
#> 3  fixed                V:N    TRUE    TRUE
#> 4 random    vsc( isc( B ) )   FALSE   FALSE
#> 5 random vsc( isc( B:MP ) )   FALSE   FALSE

pp=predict(object=m3, Dtable=Dt, D="N")
pp$pvals
#>       N predicted.value std.error
#> 0     0        79.38889  7.153894
#> 0.2 0.2        98.88889  7.153894
#> 0.4 0.4       114.22222  7.153894
#> 0.6 0.6       123.38889  7.153894

#############################
## predict means for variety
#############################

Dt <- m3$Dtable; Dt
#>     type               term include average
#> 1  fixed                  V   FALSE   FALSE
#> 2  fixed                  N   FALSE   FALSE
#> 3  fixed                V:N   FALSE   FALSE
#> 4 random    vsc( isc( B ) )   FALSE   FALSE
#> 5 random vsc( isc( B:MP ) )   FALSE   FALSE
# first fixed effect include
Dt[1,"include"] = TRUE
# second fixed effect just average
Dt[2,"average"] = TRUE
# third fixed effect include and average
Dt[3,"include"] = TRUE
Dt[3,"average"] = TRUE
Dt
#>     type               term include average
#> 1  fixed                  V    TRUE   FALSE
#> 2  fixed                  N   FALSE    TRUE
#> 3  fixed                V:N    TRUE    TRUE
#> 4 random    vsc( isc( B ) )   FALSE   FALSE
#> 5 random vsc( isc( B:MP ) )   FALSE   FALSE

pp=predict(object=m3, Dtable=Dt, D="V")
pp$pvals
#>                     V predicted.value std.error
#> GoldenRain GoldenRain        104.5000  7.777895
#> Marvellous Marvellous        109.7917  7.777895
#> ictory         ictory         97.6250  7.777895

#############################
## predict means for nitrogen:variety
#############################
# prediction matrix D based on (equivalent to classify in asreml)
Dt <- m3$Dtable; Dt
#>     type               term include average
#> 1  fixed                  V   FALSE   FALSE
#> 2  fixed                  N   FALSE   FALSE
#> 3  fixed                V:N   FALSE   FALSE
#> 4 random    vsc( isc( B ) )   FALSE   FALSE
#> 5 random vsc( isc( B:MP ) )   FALSE   FALSE
# first fixed effect include and average
Dt[1,"include"] = TRUE
Dt[1,"average"] = TRUE
# second fixed effect include and average
Dt[2,"include"] = TRUE
Dt[2,"average"] = TRUE
# third fixed effect include and average
Dt[3,"include"] = TRUE
Dt[3,"average"] = TRUE
Dt
#>     type               term include average
#> 1  fixed                  V    TRUE    TRUE
#> 2  fixed                  N    TRUE    TRUE
#> 3  fixed                V:N    TRUE    TRUE
#> 4 random    vsc( isc( B ) )   FALSE   FALSE
#> 5 random vsc( isc( B:MP ) )   FALSE   FALSE

pp=predict(object=m3, Dtable=Dt, D="N:V")
pp$pvals
#>                               N:V predicted.value std.error
#> N0:VGoldenRain     N0:VGoldenRain        80.00000  9.091182
#> N0.2:VGoldenRain N0.2:VGoldenRain        89.25000  8.238950
#> N0.4:VGoldenRain N0.4:VGoldenRain        97.33333  8.238950
#> N0.6:VGoldenRain N0.6:VGoldenRain       102.41667  8.238950
#> N0:VMarvellous     N0:VMarvellous        83.33333  7.684143
#> N0.2:VMarvellous N0.2:VMarvellous        94.25000  7.684143
#> N0.4:VMarvellous N0.4:VMarvellous        98.58333  7.684143
#> N0.6:VMarvellous N0.6:VMarvellous       103.41667  7.684143
#> N0:VVictory           N0:VVictory        75.75000  7.684143
#> N0.2:VVictory       N0.2:VVictory        84.83333  7.684143
#> N0.4:VVictory       N0.4:VVictory        95.41667  7.684143
#> N0.6:VVictory       N0.6:VVictory        99.25000  7.684143