Predict form of a LMM fitted with mmer

predict method for class "mmer".

Usage

# S3 method for class 'mmer'
predict(object, Dtable=NULL, D, ...)

Arguments

object

a mixed model of class "mmer"

Dtable

a table specifying the terms to be included or averaged.

An "include" term means that the model matrices for that fixed or random effect is filled with 1's for the positions where column names and row names match.

An "include and average" term means that the model matrices for that fixed or random effect is filled with 1/#1's in that row.

An "average" term alone means that all rows for such fixed or random effect will be filled with 1/#levels in the effect.

If a term is not considered "include" or "average" is then totally ignored in the BLUP and SE calculation.

The default rule to invoke when the user doesn't provide the Dtable is to include and average all terms that match the argument D.

D

a character string specifying the variable used to extract levels for the rows of the D matrix and its construction. Alternatively, the D matrix (of class dgCMatrix) specifying the matrix to be used for the predictions directly.

...

Further arguments to be passed.

Details

This function allows to produce predictions specifying those variables that define the margins of the hypertable to be predicted (argument D). Predictions are obtained for each combination of values of the specified variables that is present in the data set used to fit the model. See vignettes for more details.

For predicted values the pertinent design matrices X and Z together with BLUEs (b) and BLUPs (u) are multiplied and added together.

predicted.value equal Xb + Zu.1 + ... + Zu.n

For computing standard errors for predictions the parts of the coefficient matrix:

C11 equal (X.t() V.inv() X).inv()

C12 equal 0 - [(X.t() V.inv() X).inv() X.t() V.inv() G Z]

C22 equal PEV equal G - [Z.t() G[V.inv() - (V.inv() X X.t() V.inv() X V.inv() X)]G Z.t()]

In practive C equals ( W.t() V.inv() W ).inv()

when both fixed and random effects are present in the inclusion set. If only fixed and random effects are included, only the respective terms from the SE for fixed or random effects are calculated.

Value

pvals

the table of predictions according to the specified arguments.

vcov

the variance covariance for the predictions.

D

the model matrix for predictions as defined in Welham et al.(2004).

Dtable

the table specifying the terms to include and terms to be averaged.

References

Welham, S., Cullis, B., Gogel, B., Gilmour, A., and Thompson, R. (2004). Prediction in linear mixed models. Australian and New Zealand Journal of Statistics, 46, 325 - 347.

Author

Giovanny Covarrubias

See also

Examples

data(DT_yatesoats)
DT <- DT_yatesoats
m3 <- mmer(fixed=Y ~ V + N + V:N ,
           random = ~ B + B:MP,
           rcov=~units,
           data = DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -4.46709   1:16:21      0           0
#>     2      -1.34644   1:16:21      0           0
#>     3      -0.788146   1:16:21      0           0
#>     4      -0.746308   1:16:21      0           0
#>     5      -0.745925   1:16:21      0           0

#############################
## predict means for nitrogen
#############################
pp=predict(object=m3, D="N")
pp$pvals
#>       N predicted.value std.error
#> 0     0         80.0000  9.106654
#> 0.2 0.2         98.5000  9.106654
#> 0.4 0.4        114.6667  9.106654
#> 0.6 0.6        124.8333  9.106654

#############################
## predict means for variety
#############################
pp=predict(object=m3, D="V")
pp$pvals
#>                     V predicted.value std.error
#> GoldenRain GoldenRain        80.00000  9.106654
#> Marvellous Marvellous        86.66667  9.106654
#> ictory         ictory        71.50000  9.106654