Create a GE correlation matrix for simulation purposes.

Makes a simple correlation matrix based on the number of environments and megaenvironments desired.

Usage

simGECorMat(nEnv,nMegaEnv,mu=0.7, v=0.2, mu2=0, v2=0.3)

Arguments

nEnv

Number of environments to simulate. Needs to be divisible by the nMegaEnv argument.

nMegaEnv

Number of megaenvironments to simulate.

mu

Mean value of the genetic correlation within megaenvironments.

v

variance in the genetic correlation within megaenvironments.

mu2

Mean value of the genetic correlation between megaenvironments.

v2

variance in the genetic correlation between megaenvironments.

Details

Simple simulation of a correlation matrix for environments and megaenvironments.

Value

G the correlation matrix

$G

the correlation matrix

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

See also

Examples

simGECorMat(9,3)
#>              [,1]         [,2]        [,3]         [,4]       [,5]        [,6]
#>  [1,]  1.00000000  0.896955994  0.45505242  0.068243833  0.4906042  0.10349851
#>  [2,]  0.89695599  1.000000000  0.84194524 -0.008220421  0.1313082  0.13031876
#>  [3,]  0.45505242  0.841945244  1.00000000  0.060263337 -0.3580608  0.39104330
#>  [4,]  0.06824383 -0.008220421  0.06026334  1.000000000  0.6781560  0.98000000
#>  [5,]  0.49060424  0.131308166 -0.35806076  0.678156001  1.0000000  0.65131106
#>  [6,]  0.10349851  0.130318758  0.39104330  0.980000000  0.6513111  1.00000000
#>  [7,] -0.33813986 -0.554691093  0.05865872  0.180469568  0.1917474 -0.41699785
#>  [8,] -0.33890336 -0.342346258 -0.20660145  0.051059345  0.1783450 -0.24450980
#>  [9,] -0.85930405  0.286147045  0.18458543 -0.210261210 -0.3220984 -0.08999736
#>              [,7]        [,8]        [,9]
#>  [1,] -0.33813986 -0.33890336 -0.85930405
#>  [2,] -0.55469109 -0.34234626  0.28614704
#>  [3,]  0.05865872 -0.20660145  0.18458543
#>  [4,]  0.18046957  0.05105935 -0.21026121
#>  [5,]  0.19174745  0.17834503 -0.32209840
#>  [6,] -0.41699785 -0.24450980 -0.08999736
#>  [7,]  1.00000000  0.39457860  0.79836687
#>  [8,]  0.39457860  1.00000000  0.77090073
#>  [9,]  0.79836687  0.77090073  1.00000000