Stacking traits in a dataset

stackTrait creates a dataset stacking traits in the long format to be used with the mmec solver for multi-trait models.

Usage

stackTrait(data, traits)

Arguments

data

a data frame with traits in wide format.

traits

variable names corresponding to the traits that should be in the long format.

Value

$res

a data frame with traits in long format.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Author

Giovanny Covarrubias-Pazaran

See also

Examples

data(DT_example)
DT <- DT_example
A <- A_example
head(DT)
#>                   Name     Env Loc Year     Block Yield    Weight
#> 33  Manistee(MSL292-A) CA.2013  CA 2013 CA.2013.1     4 -1.904711
#> 65          CO02024-9W CA.2013  CA 2013 CA.2013.1     5 -1.446958
#> 66  Manistee(MSL292-A) CA.2013  CA 2013 CA.2013.2     5 -1.516271
#> 67            MSL007-B CA.2011  CA 2011 CA.2011.2     5 -1.435510
#> 68           MSR169-8Y CA.2013  CA 2013 CA.2013.1     5 -1.469051
#> 103         AC05153-1W CA.2013  CA 2013 CA.2013.1     6 -1.307167

DT2 <- stackTrait(DT, traits = c("Yield","Weight"))
head(DT2)
#> $long
#>                     Name     Env Loc Year     Block  trait        value
#> 1     Manistee(MSL292-A) CA.2013  CA 2013 CA.2013.1  Yield  4.000000000
#> 2             CO02024-9W CA.2013  CA 2013 CA.2013.1  Yield  5.000000000
#> 3     Manistee(MSL292-A) CA.2013  CA 2013 CA.2013.2  Yield  5.000000000
#> 4               MSL007-B CA.2011  CA 2011 CA.2011.2  Yield  5.000000000
#> 5              MSR169-8Y CA.2013  CA 2013 CA.2013.1  Yield  5.000000000
#> 6             AC05153-1W CA.2013  CA 2013 CA.2013.1  Yield  6.000000000
#> 7               AF4157-6 CA.2012  CA 2012 CA.2012.2  Yield  6.000000000
#> 8               MSL007-B CA.2011  CA 2011 CA.2011.1  Yield  6.000000000
#> 9              MSR128-4Y CA.2012  CA 2012 CA.2012.1  Yield  6.000000000
#> 10        NY153(NYH25-4) CA.2012  CA 2012 CA.2012.2  Yield  6.000000000
#> 11        NY153(NYH25-4) CA.2012  CA 2012 CA.2012.1  Yield  6.000000000
#> 12               W8822-1 CA.2013  CA 2013 CA.2013.1  Yield  6.000000000
#> 13            AC05153-1W CA.2013  CA 2013 CA.2013.2  Yield  7.000000000
#> 14            CO02024-9W CA.2013  CA 2013 CA.2013.2  Yield  7.000000000
#> 15            CO05061-6W CA.2013  CA 2013 CA.2013.1  Yield  7.000000000
#> 16              MSK061-4 CA.2012  CA 2012 CA.2012.1  Yield  7.000000000
#> 17              MSL007-B CA.2013  CA 2013 CA.2013.2  Yield  7.000000000
#> 18              MSL007-B CA.2013  CA 2013 CA.2013.1  Yield  7.000000000
#> 19              MSM246-B CA.2013  CA 2013 CA.2013.2  Yield  7.000000000
#> 20              MSM246-B CA.2013  CA 2013 CA.2013.1  Yield  7.000000000
#> 21              MSR058-1 CA.2012  CA 2012 CA.2012.1  Yield  7.000000000
#> 22              MSR127-2 CA.2012  CA 2012 CA.2012.2  Yield  7.000000000
#> 23             MSR128-4Y CA.2013  CA 2013 CA.2013.1  Yield  7.000000000
#> 24             MSR169-8Y CA.2012  CA 2012 CA.2012.2  Yield  7.000000000
#> 25             MSR169-8Y CA.2013  CA 2013 CA.2013.2  Yield  7.000000000
#> 26        NY153(NYH25-4) CA.2013  CA 2013 CA.2013.1  Yield  7.000000000
#> 27               W8822-1 CA.2013  CA 2013 CA.2013.2  Yield  7.000000000
#> 28               W8822-2 CA.2013  CA 2013 CA.2013.2  Yield  7.000000000
#> 29               W8867-5 CA.2012  CA 2012 CA.2012.2  Yield  7.000000000
#> 30               W8867-5 CA.2012  CA 2012 CA.2012.1  Yield  7.000000000
#> 31             A01143-3C CA.2012  CA 2012 CA.2012.1  Yield  8.000000000
#> 32            AC00206-2W CA.2013  CA 2013 CA.2013.2  Yield  8.000000000
#> 33            AC03433-1W CA.2013  CA 2013 CA.2013.1  Yield  8.000000000
#> 34            AC05153-1W CA.2012  CA 2012 CA.2012.2  Yield  8.000000000
#> 35              AF4157-6 CA.2012  CA 2012 CA.2012.1  Yield  8.000000000
#> 36              AF4157-6 CA.2013  CA 2013 CA.2013.2  Yield  8.000000000
#> 37              AF4157-6 CA.2013  CA 2013 CA.2013.1  Yield  8.000000000
#> 38              AF4648-2 CA.2012  CA 2012 CA.2012.2  Yield  8.000000000
#> 39              AF4648-2 CA.2012  CA 2012 CA.2012.1  Yield  8.000000000
#> 40            CO05061-2P CA.2012  CA 2012 CA.2012.1  Yield  8.000000000
#> 41            CO05061-6W CA.2013  CA 2013 CA.2013.2  Yield  8.000000000
#> 42    Manistee(MSL292-A) CA.2012  CA 2012 CA.2012.2  Yield  8.000000000
#> 43    Manistee(MSL292-A) CA.2012  CA 2012 CA.2012.1  Yield  8.000000000
#> 44              MSR057-4 CA.2013  CA 2013 CA.2013.2  Yield  8.000000000
#> 45              MSR058-1 CA.2012  CA 2012 CA.2012.2  Yield  8.000000000
#> 46             MSR128-4Y CA.2012  CA 2012 CA.2012.2  Yield  8.000000000
#> 47             MSR128-4Y CA.2013  CA 2013 CA.2013.2  Yield  8.000000000
#> 48             MSR157-1Y CA.2012  CA 2012 CA.2012.1  Yield  8.000000000
#> 49            AC01151-5W CA.2012  CA 2012 CA.2012.2  Yield  9.000000000
#> 50            AC03433-1W CA.2013  CA 2013 CA.2013.2  Yield  9.000000000
#> 51            AC05153-1W CA.2012  CA 2012 CA.2012.1  Yield  9.000000000
#> 52  Accumulator(W2324-1) CA.2013  CA 2013 CA.2013.1  Yield  9.000000000
#> 53              MSK061-4 CA.2011  CA 2011 CA.2011.2  Yield  9.000000000
#> 54              MSK061-4 CA.2012  CA 2012 CA.2012.2  Yield  9.000000000
#> 55              MSK061-4 CA.2013  CA 2013 CA.2013.1  Yield  9.000000000
#> 56              MSL007-B CA.2012  CA 2012 CA.2012.2  Yield  9.000000000
#> 57              MSL007-B CA.2012  CA 2012 CA.2012.1  Yield  9.000000000
#> 58              MSQ086-3 CA.2013  CA 2013 CA.2013.2  Yield  9.000000000
#> 59              MSQ086-3 CA.2013  CA 2013 CA.2013.1  Yield  9.000000000
#> 60              MSR057-4 CA.2013  CA 2013 CA.2013.1  Yield  9.000000000
#> 61             MSR157-1Y CA.2012  CA 2012 CA.2012.2  Yield  9.000000000
#> 62             MSS165-2Y CA.2012  CA 2012 CA.2012.1  Yield  9.000000000
#> 63                 NY148 CA.2012  CA 2012 CA.2012.1  Yield  9.000000000
#> 64              W8539-2Y CA.2012  CA 2012 CA.2012.2  Yield  9.000000000
#> 65               W8822-2 CA.2013  CA 2013 CA.2013.1  Yield  9.000000000
#> 66               W8822-3 CA.2012  CA 2012 CA.2012.1  Yield  9.000000000
#> 67               W8822-3 CA.2013  CA 2013 CA.2013.2  Yield  9.000000000
#> 68               W8867-7 CA.2013  CA 2013 CA.2013.1  Yield  9.000000000
#> 69            AC00206-2W CA.2012  CA 2012 CA.2012.1  Yield 10.000000000
#> 70            AC01151-5W CA.2012  CA 2012 CA.2012.1  Yield 10.000000000
#> 71            AC03433-1W CA.2012  CA 2012 CA.2012.1  Yield 10.000000000
#> 72              Atlantic CA.2012  CA 2012 CA.2012.1  Yield 10.000000000
#> 73            CO02033-1W CA.2012  CA 2012 CA.2012.2  Yield 10.000000000
#> 74            CO02033-1W CA.2013  CA 2013 CA.2013.1  Yield 10.000000000
#> 75            CO05061-2P CA.2012  CA 2012 CA.2012.2  Yield 10.000000000
#> 76            CO05061-6W CA.2012  CA 2012 CA.2012.1  Yield 10.000000000
#> 77              MSK061-4 CA.2011  CA 2011 CA.2011.1  Yield 10.000000000
#> 78              MSR058-1 CA.2013  CA 2013 CA.2013.2  Yield 10.000000000
#> 79              MSR058-1 CA.2013  CA 2013 CA.2013.1  Yield 10.000000000
#> 80             MSR169-8Y CA.2012  CA 2012 CA.2012.1  Yield 10.000000000
#> 81        NY153(NYH25-4) CA.2013  CA 2013 CA.2013.2  Yield 10.000000000
#> 82               W8615-5 CA.2012  CA 2012 CA.2012.1  Yield 10.000000000
#> 83               W8867-7 CA.2012  CA 2012 CA.2012.1  Yield 10.000000000
#> 84             A01143-3C CA.2012  CA 2012 CA.2012.2  Yield 11.000000000
#> 85            AC00206-2W CA.2012  CA 2012 CA.2012.2  Yield 11.000000000
#> 86            AC00206-2W CA.2013  CA 2013 CA.2013.1  Yield 11.000000000
#> 87            AC03433-1W CA.2012  CA 2012 CA.2012.2  Yield 11.000000000
#> 88            AC03452-2W CA.2013  CA 2013 CA.2013.2  Yield 11.000000000
#> 89              Atlantic CA.2012  CA 2012 CA.2012.2  Yield 11.000000000
#> 90              Atlantic CA.2013  CA 2013 CA.2013.2  Yield 11.000000000
#> 91            CO02024-9W CA.2012  CA 2012 CA.2012.2  Yield 11.000000000
#> 92            CO02024-9W CA.2012  CA 2012 CA.2012.1  Yield 11.000000000
#> 93            CO02033-1W CA.2012  CA 2012 CA.2012.1  Yield 11.000000000
#> 94            CO02033-1W CA.2013  CA 2013 CA.2013.2  Yield 11.000000000
#> 95              MSQ089-1 CA.2012  CA 2012 CA.2012.2  Yield 11.000000000
#> 96              MSQ089-1 CA.2013  CA 2013 CA.2013.2  Yield 11.000000000
#> 97              MSR127-2 CA.2013  CA 2013 CA.2013.1  Yield 11.000000000
#> 98             MSR169-8Y CA.2011  CA 2011 CA.2011.1  Yield 11.000000000
#> 99             MSS165-2Y CA.2012  CA 2012 CA.2012.2  Yield 11.000000000
#> 100                NY148 CA.2011  CA 2011 CA.2011.1  Yield 11.000000000
#> 101              W8822-1 CA.2012  CA 2012 CA.2012.2  Yield 11.000000000
#> 102              W8822-3 CA.2013  CA 2013 CA.2013.1  Yield 11.000000000
#> 103              W8867-5 CA.2013  CA 2013 CA.2013.1  Yield 11.000000000
#> 104              W8867-7 CA.2013  CA 2013 CA.2013.2  Yield 11.000000000
#> 105             AF4648-2 CA.2013  CA 2013 CA.2013.1  Yield 12.000000000
#> 106           CO05061-6W CA.2012  CA 2012 CA.2012.2  Yield 12.000000000
#> 107             MSK061-4 CA.2013  CA 2013 CA.2013.2  Yield 12.000000000
#> 108             MSQ086-3 CA.2012  CA 2012 CA.2012.2  Yield 12.000000000
#> 109             MSQ089-1 CA.2012  CA 2012 CA.2012.1  Yield 12.000000000
#> 110             MSR148-4 CA.2012  CA 2012 CA.2012.1  Yield 12.000000000
#> 111            MSR157-1Y CA.2013  CA 2013 CA.2013.1  Yield 12.000000000
#> 112            MSR169-8Y CA.2011  CA 2011 CA.2011.2  Yield 12.000000000
#> 113                NY148 CA.2012  CA 2012 CA.2012.2  Yield 12.000000000
#> 114       NY152(NYH15-5) CA.2012  CA 2012 CA.2012.2  Yield 12.000000000
#> 115       NY152(NYH15-5) CA.2012  CA 2012 CA.2012.1  Yield 12.000000000
#> 116              Snowden CA.2013  CA 2013 CA.2013.2  Yield 12.000000000
#> 117             W8539-2Y CA.2012  CA 2012 CA.2012.1  Yield 12.000000000
#> 118              W8867-7 CA.2012  CA 2012 CA.2012.2  Yield 12.000000000
#> 119           AC03433-1W CA.2011  CA 2011 CA.2011.2  Yield 13.000000000
#> 120             AF4157-6 CA.2011  CA 2011 CA.2011.1  Yield 13.000000000
#> 121             Atlantic CA.2013  CA 2013 CA.2013.1  Yield 13.000000000
#> 122           CO00270-7W CA.2012  CA 2012 CA.2012.1  Yield 13.000000000
#> 123             MSM246-B CA.2011  CA 2011 CA.2011.1  Yield 13.000000000
#> 124             MSM246-B CA.2012  CA 2012 CA.2012.2  Yield 13.000000000
#> 125             MSM246-B CA.2012  CA 2012 CA.2012.1  Yield 13.000000000
#> 126             MSQ086-3 CA.2012  CA 2012 CA.2012.1  Yield 13.000000000
#> 127             MSR057-4 CA.2012  CA 2012 CA.2012.2  Yield 13.000000000
#> 128             MSR057-4 CA.2012  CA 2012 CA.2012.1  Yield 13.000000000
#> 129             MSR127-2 CA.2013  CA 2013 CA.2013.2  Yield 13.000000000
#> 130              Snowden CA.2013  CA 2013 CA.2013.1  Yield 13.000000000
#> 131              W8822-3 CA.2012  CA 2012 CA.2012.2  Yield 13.000000000
#> 132              W8867-5 CA.2013  CA 2013 CA.2013.2  Yield 13.000000000
#> 133           AC03433-1W CA.2011  CA 2011 CA.2011.1  Yield 14.000000000
#> 134           AC03452-2W CA.2012  CA 2012 CA.2012.2  Yield 14.000000000
#> 135           AC03452-2W CA.2012  CA 2012 CA.2012.1  Yield 14.000000000
#> 136           AC03452-2W CA.2013  CA 2013 CA.2013.1  Yield 14.000000000
#> 137             AF4648-2 CA.2013  CA 2013 CA.2013.2  Yield 14.000000000
#> 138             MSQ089-1 CA.2013  CA 2013 CA.2013.1  Yield 14.000000000
#> 139             MSR148-4 CA.2012  CA 2012 CA.2012.2  Yield 14.000000000
#> 140       NY152(NYH15-5) CA.2013  CA 2013 CA.2013.2  Yield 14.000000000
#> 141              W8603-1 CA.2013  CA 2013 CA.2013.1  Yield 14.000000000
#> 142              W8822-1 CA.2012  CA 2012 CA.2012.1  Yield 14.000000000
#> 143 Accumulator(W2324-1) CA.2013  CA 2013 CA.2013.2  Yield 15.000000000
#> 144             Atlantic CA.2011  CA 2011 CA.2011.2  Yield 15.000000000
#> 145             Atlantic CA.2011  CA 2011 CA.2011.1  Yield 15.000000000
#> 146           CO00270-7W CA.2012  CA 2012 CA.2012.2  Yield 15.000000000
#> 147             MSR058-1 CA.2011  CA 2011 CA.2011.1  Yield 15.000000000
#> 148            MSR128-4Y CA.2011  CA 2011 CA.2011.1  Yield 15.000000000
#> 149                NY148 CA.2011  CA 2011 CA.2011.2  Yield 15.000000000
#> 150       NY152(NYH15-5) CA.2013  CA 2013 CA.2013.1  Yield 15.000000000
#> 151              W8603-1 CA.2013  CA 2013 CA.2013.2  Yield 15.000000000
#> 152 Accumulator(W2324-1) CA.2012  CA 2012 CA.2012.2  Yield 16.000000000
#> 153             MSQ089-1 CA.2011  CA 2011 CA.2011.1  Yield 16.000000000
#> 154             MSR058-1 CA.2011  CA 2011 CA.2011.2  Yield 16.000000000
#> 155             MSR127-2 CA.2012  CA 2012 CA.2012.1  Yield 16.000000000
#> 156              W8603-1 CA.2011  CA 2011 CA.2011.2  Yield 16.000000000
#> 157            A01143-3C CA.2013  CA 2013 CA.2013.2  Yield 17.000000000
#> 158             AF4157-6 CA.2011  CA 2011 CA.2011.2  Yield 17.000000000
#> 159           CO02024-9W CA.2011  CA 2011 CA.2011.1  Yield 17.000000000
#> 160             MSR148-4 CA.2011  CA 2011 CA.2011.2  Yield 17.000000000
#> 161             W8539-2Y CA.2011  CA 2011 CA.2011.2  Yield 17.000000000
#> 162              W8615-5 CA.2011  CA 2011 CA.2011.2  Yield 17.000000000
#> 163           AC01151-5W CA.2011  CA 2011 CA.2011.2  Yield 18.000000000
#> 164 Accumulator(W2324-1) CA.2012  CA 2012 CA.2012.1  Yield 18.000000000
#> 165           CO02024-9W CA.2011  CA 2011 CA.2011.2  Yield 18.000000000
#> 166             MSM246-B CA.2011  CA 2011 CA.2011.2  Yield 18.000000000
#> 167              W8603-1 CA.2011  CA 2011 CA.2011.1  Yield 18.000000000
#> 168              W8822-2 CA.2012  CA 2012 CA.2012.1  Yield 18.000000000
#> 169           CO00270-7W CA.2011  CA 2011 CA.2011.2  Yield 19.000000000
#> 170           CO00270-7W CA.2011  CA 2011 CA.2011.1  Yield 19.000000000
#> 171             MSQ089-1 CA.2011  CA 2011 CA.2011.2  Yield 19.000000000
#> 172            MSS165-2Y CA.2011  CA 2011 CA.2011.2  Yield 19.000000000
#> 173             W8539-2Y CA.2011  CA 2011 CA.2011.1  Yield 19.000000000
#> 174             MSR127-2 CA.2011  CA 2011 CA.2011.2  Yield 20.000000000
#> 175            MSS165-2Y CA.2011  CA 2011 CA.2011.1  Yield 20.000000000
#> 176             MSR127-2 CA.2011  CA 2011 CA.2011.1  Yield 21.000000000
#> 177              W8615-5 CA.2011  CA 2011 CA.2011.1  Yield 21.000000000
#> 178            A01143-3C CA.2011  CA 2011 CA.2011.1  Yield 22.000000000
#> 179 Accumulator(W2324-1) CA.2011  CA 2011 CA.2011.1  Yield 22.000000000
#> 180            A01143-3C CA.2011  CA 2011 CA.2011.2  Yield 23.000000000
#> 181            A01143-3C CA.2013  CA 2013 CA.2013.1  Yield 23.000000000
#> 182           AC01151-5W CA.2011  CA 2011 CA.2011.1  Yield 23.000000000
#> 183             MSR148-4 CA.2011  CA 2011 CA.2011.1  Yield 23.000000000
#> 184              W8615-5 CA.2012  CA 2012 CA.2012.2  Yield 25.000000000
#> 185 Accumulator(W2324-1) CA.2011  CA 2011 CA.2011.2  Yield 29.000000000
#> 186   Manistee(MSL292-A) CA.2013  CA 2013 CA.2013.1 Weight -1.904710944
#> 187           CO02024-9W CA.2013  CA 2013 CA.2013.1 Weight -1.446958222
#> 188   Manistee(MSL292-A) CA.2013  CA 2013 CA.2013.2 Weight -1.516271359
#> 189             MSL007-B CA.2011  CA 2011 CA.2011.2 Weight -1.435510310
#> 190            MSR169-8Y CA.2013  CA 2013 CA.2013.1 Weight -1.469051308
#> 191           AC05153-1W CA.2013  CA 2013 CA.2013.1 Weight -1.307167384
#> 192             AF4157-6 CA.2012  CA 2012 CA.2012.2 Weight -1.252885127
#> 193             MSL007-B CA.2011  CA 2011 CA.2011.1 Weight -1.438962493
#> 194            MSR128-4Y CA.2012  CA 2012 CA.2012.1 Weight -1.362001169
#> 195       NY153(NYH25-4) CA.2012  CA 2012 CA.2012.2 Weight -1.180293079
#> 196       NY153(NYH25-4) CA.2012  CA 2012 CA.2012.1 Weight -0.686940586
#> 197              W8822-1 CA.2013  CA 2013 CA.2013.1 Weight -0.972147457
#> 198           AC05153-1W CA.2013  CA 2013 CA.2013.2 Weight -1.076444843
#> 199           CO02024-9W CA.2013  CA 2013 CA.2013.2 Weight -1.159478980
#> 200           CO05061-6W CA.2013  CA 2013 CA.2013.1 Weight -1.188148689
#> 201             MSK061-4 CA.2012  CA 2012 CA.2012.1 Weight -1.246607923
#> 202             MSL007-B CA.2013  CA 2013 CA.2013.2 Weight -1.145567730
#> 203             MSL007-B CA.2013  CA 2013 CA.2013.1 Weight -0.819852994
#> 204             MSM246-B CA.2013  CA 2013 CA.2013.2 Weight -1.251570337
#> 205             MSM246-B CA.2013  CA 2013 CA.2013.1 Weight -1.219755634
#> 206             MSR058-1 CA.2012  CA 2012 CA.2012.1 Weight -1.017792171
#> 207             MSR127-2 CA.2012  CA 2012 CA.2012.2 Weight -0.671175400
#> 208            MSR128-4Y CA.2013  CA 2013 CA.2013.1 Weight -1.111328466
#> 209            MSR169-8Y CA.2012  CA 2012 CA.2012.2 Weight -1.283317496
#> 210            MSR169-8Y CA.2013  CA 2013 CA.2013.2 Weight -1.020208579
#> 211       NY153(NYH25-4) CA.2013  CA 2013 CA.2013.1 Weight -1.081052481
#> 212              W8822-1 CA.2013  CA 2013 CA.2013.2 Weight -0.867852941
#> 213              W8822-2 CA.2013  CA 2013 CA.2013.2 Weight -1.234828942
#> 214              W8867-5 CA.2012  CA 2012 CA.2012.2 Weight -0.429062666
#> 215              W8867-5 CA.2012  CA 2012 CA.2012.1 Weight -1.195027706
#> 216            A01143-3C CA.2012  CA 2012 CA.2012.1 Weight -0.589984373
#> 217           AC00206-2W CA.2013  CA 2013 CA.2013.2 Weight -0.323892485
#> 218           AC03433-1W CA.2013  CA 2013 CA.2013.1 Weight -1.226764427
#> 219           AC05153-1W CA.2012  CA 2012 CA.2012.2 Weight -0.655897507
#> 220             AF4157-6 CA.2012  CA 2012 CA.2012.1 Weight -0.587283518
#> 221             AF4157-6 CA.2013  CA 2013 CA.2013.2 Weight -0.636425366
#> 222             AF4157-6 CA.2013  CA 2013 CA.2013.1 Weight -0.879965476
#> 223             AF4648-2 CA.2012  CA 2012 CA.2012.2 Weight -0.448960197
#> 224             AF4648-2 CA.2012  CA 2012 CA.2012.1 Weight -0.659728628
#> 225           CO05061-2P CA.2012  CA 2012 CA.2012.1 Weight -0.958733815
#> 226           CO05061-6W CA.2013  CA 2013 CA.2013.2 Weight -0.122812023
#> 227   Manistee(MSL292-A) CA.2012  CA 2012 CA.2012.2 Weight -1.043605282
#> 228   Manistee(MSL292-A) CA.2012  CA 2012 CA.2012.1 Weight -0.828186933
#> 229             MSR057-4 CA.2013  CA 2013 CA.2013.2 Weight -0.854903222
#> 230             MSR058-1 CA.2012  CA 2012 CA.2012.2 Weight -0.318770439
#> 231            MSR128-4Y CA.2012  CA 2012 CA.2012.2 Weight -0.488842618
#> 232            MSR128-4Y CA.2013  CA 2013 CA.2013.2 Weight -0.735799397
#> 233            MSR157-1Y CA.2012  CA 2012 CA.2012.1 Weight -0.865993142
#> 234           AC01151-5W CA.2012  CA 2012 CA.2012.2 Weight -0.628242662
#> 235           AC03433-1W CA.2013  CA 2013 CA.2013.2 Weight -0.936532122
#> 236           AC05153-1W CA.2012  CA 2012 CA.2012.1 Weight -0.594419352
#> 237 Accumulator(W2324-1) CA.2013  CA 2013 CA.2013.1 Weight -0.247575971
#> 238             MSK061-4 CA.2011  CA 2011 CA.2011.2 Weight -0.798393744
#> 239             MSK061-4 CA.2012  CA 2012 CA.2012.2 Weight -0.788407168
#> 240             MSK061-4 CA.2013  CA 2013 CA.2013.1 Weight -0.705221520
#> 241             MSL007-B CA.2012  CA 2012 CA.2012.2 Weight -0.348264860
#> 242             MSL007-B CA.2012  CA 2012 CA.2012.1 Weight -0.338483315
#> 243             MSQ086-3 CA.2013  CA 2013 CA.2013.2 Weight -0.755715811
#> 244             MSQ086-3 CA.2013  CA 2013 CA.2013.1 Weight -0.752285772
#> 245             MSR057-4 CA.2013  CA 2013 CA.2013.1 Weight -0.735646645
#> 246            MSR157-1Y CA.2012  CA 2012 CA.2012.2 Weight -0.842713637
#> 247            MSS165-2Y CA.2012  CA 2012 CA.2012.1 Weight -0.697167716
#> 248                NY148 CA.2012  CA 2012 CA.2012.1 Weight -0.955954922
#> 249             W8539-2Y CA.2012  CA 2012 CA.2012.2 Weight -0.634630752
#> 250              W8822-2 CA.2013  CA 2013 CA.2013.1 Weight -1.031659120
#> 251              W8822-3 CA.2012  CA 2012 CA.2012.1 Weight -0.522481207
#> 252              W8822-3 CA.2013  CA 2013 CA.2013.2 Weight -1.062704793
#> 253              W8867-7 CA.2013  CA 2013 CA.2013.1 Weight -0.746255681
#> 254           AC00206-2W CA.2012  CA 2012 CA.2012.1 Weight -0.546320423
#> 255           AC01151-5W CA.2012  CA 2012 CA.2012.1 Weight -0.116574612
#> 256           AC03433-1W CA.2012  CA 2012 CA.2012.1 Weight -0.671755443
#> 257             Atlantic CA.2012  CA 2012 CA.2012.1 Weight -0.338181708
#> 258           CO02033-1W CA.2012  CA 2012 CA.2012.2 Weight -0.412168564
#> 259           CO02033-1W CA.2013  CA 2013 CA.2013.1 Weight -0.495521488
#> 260           CO05061-2P CA.2012  CA 2012 CA.2012.2 Weight -0.057343421
#> 261           CO05061-6W CA.2012  CA 2012 CA.2012.1 Weight -0.248603108
#> 262             MSK061-4 CA.2011  CA 2011 CA.2011.1 Weight -0.190226553
#> 263             MSR058-1 CA.2013  CA 2013 CA.2013.2 Weight -0.538487643
#> 264             MSR058-1 CA.2013  CA 2013 CA.2013.1 Weight -0.264204088
#> 265            MSR169-8Y CA.2012  CA 2012 CA.2012.1 Weight -0.463717649
#> 266       NY153(NYH25-4) CA.2013  CA 2013 CA.2013.2 Weight -0.833799536
#> 267              W8615-5 CA.2012  CA 2012 CA.2012.1 Weight -0.103570522
#> 268              W8867-7 CA.2012  CA 2012 CA.2012.1 Weight -0.568484395
#> 269            A01143-3C CA.2012  CA 2012 CA.2012.2 Weight  0.098566285
#> 270           AC00206-2W CA.2012  CA 2012 CA.2012.2 Weight  0.066715229
#> 271           AC00206-2W CA.2013  CA 2013 CA.2013.1 Weight -0.524833610
#> 272           AC03433-1W CA.2012  CA 2012 CA.2012.2 Weight -0.325822019
#> 273           AC03452-2W CA.2013  CA 2013 CA.2013.2 Weight  0.011424613
#> 274             Atlantic CA.2012  CA 2012 CA.2012.2 Weight -0.219217660
#> 275             Atlantic CA.2013  CA 2013 CA.2013.2 Weight -0.039319589
#> 276           CO02024-9W CA.2012  CA 2012 CA.2012.2 Weight  0.003720692
#> 277           CO02024-9W CA.2012  CA 2012 CA.2012.1 Weight  0.022758094
#> 278           CO02033-1W CA.2012  CA 2012 CA.2012.1 Weight -0.202947472
#> 279           CO02033-1W CA.2013  CA 2013 CA.2013.2 Weight  0.061383574
#> 280             MSQ089-1 CA.2012  CA 2012 CA.2012.2 Weight -0.136667734
#> 281             MSQ089-1 CA.2013  CA 2013 CA.2013.2 Weight -0.231594448
#> 282             MSR127-2 CA.2013  CA 2013 CA.2013.1 Weight -0.136450583
#> 283            MSR169-8Y CA.2011  CA 2011 CA.2011.1 Weight  0.026820367
#> 284            MSS165-2Y CA.2012  CA 2012 CA.2012.2 Weight -0.013424286
#> 285                NY148 CA.2011  CA 2011 CA.2011.1 Weight -0.032782056
#> 286              W8822-1 CA.2012  CA 2012 CA.2012.2 Weight -0.362160719
#> 287              W8822-3 CA.2013  CA 2013 CA.2013.1 Weight -0.330839989
#> 288              W8867-5 CA.2013  CA 2013 CA.2013.1 Weight  0.018349694
#> 289              W8867-7 CA.2013  CA 2013 CA.2013.2 Weight -0.420701351
#> 290             AF4648-2 CA.2013  CA 2013 CA.2013.1 Weight  0.192229417
#> 291           CO05061-6W CA.2012  CA 2012 CA.2012.2 Weight  0.009404862
#> 292             MSK061-4 CA.2013  CA 2013 CA.2013.2 Weight  0.056437997
#> 293             MSQ086-3 CA.2012  CA 2012 CA.2012.2 Weight -0.248137458
#> 294             MSQ089-1 CA.2012  CA 2012 CA.2012.1 Weight  0.306390867
#> 295             MSR148-4 CA.2012  CA 2012 CA.2012.1 Weight  0.161373012
#> 296            MSR157-1Y CA.2013  CA 2013 CA.2013.1 Weight  0.011815822
#> 297            MSR169-8Y CA.2011  CA 2011 CA.2011.2 Weight  0.170630337
#> 298                NY148 CA.2012  CA 2012 CA.2012.2 Weight -0.374530775
#> 299       NY152(NYH15-5) CA.2012  CA 2012 CA.2012.2 Weight -0.116689464
#> 300       NY152(NYH15-5) CA.2012  CA 2012 CA.2012.1 Weight  0.234711307
#> 301              Snowden CA.2013  CA 2013 CA.2013.2 Weight  0.144312781
#> 302             W8539-2Y CA.2012  CA 2012 CA.2012.1 Weight -0.072281390
#> 303              W8867-7 CA.2012  CA 2012 CA.2012.2 Weight  0.139732621
#> 304           AC03433-1W CA.2011  CA 2011 CA.2011.2 Weight  0.052357810
#> 305             AF4157-6 CA.2011  CA 2011 CA.2011.1 Weight  0.152735425
#> 306             Atlantic CA.2013  CA 2013 CA.2013.1 Weight  0.233158452
#> 307           CO00270-7W CA.2012  CA 2012 CA.2012.1 Weight  0.295743756
#> 308             MSM246-B CA.2011  CA 2011 CA.2011.1 Weight  0.181131304
#> 309             MSM246-B CA.2012  CA 2012 CA.2012.2 Weight  0.014473805
#> 310             MSM246-B CA.2012  CA 2012 CA.2012.1 Weight -0.123372067
#> 311             MSQ086-3 CA.2012  CA 2012 CA.2012.1 Weight -0.195697581
#> 312             MSR057-4 CA.2012  CA 2012 CA.2012.2 Weight  0.078407572
#> 313             MSR057-4 CA.2012  CA 2012 CA.2012.1 Weight  0.306855809
#> 314             MSR127-2 CA.2013  CA 2013 CA.2013.2 Weight  0.095513681
#> 315              Snowden CA.2013  CA 2013 CA.2013.1 Weight  0.323058995
#> 316              W8822-3 CA.2012  CA 2012 CA.2012.2 Weight  0.300202359
#> 317              W8867-5 CA.2013  CA 2013 CA.2013.2 Weight  0.506922239
#> 318           AC03433-1W CA.2011  CA 2011 CA.2011.1 Weight  0.541638035
#> 319           AC03452-2W CA.2012  CA 2012 CA.2012.2 Weight  0.322587942
#> 320           AC03452-2W CA.2012  CA 2012 CA.2012.1 Weight  0.229329439
#> 321           AC03452-2W CA.2013  CA 2013 CA.2013.1 Weight  0.187346766
#> 322             AF4648-2 CA.2013  CA 2013 CA.2013.2 Weight  0.058886587
#> 323             MSQ089-1 CA.2013  CA 2013 CA.2013.1 Weight  0.452509051
#> 324             MSR148-4 CA.2012  CA 2012 CA.2012.2 Weight  0.733302892
#> 325       NY152(NYH15-5) CA.2013  CA 2013 CA.2013.2 Weight  0.830753384
#> 326              W8603-1 CA.2013  CA 2013 CA.2013.1 Weight  0.755462905
#> 327              W8822-1 CA.2012  CA 2012 CA.2012.1 Weight  0.188998663
#> 328 Accumulator(W2324-1) CA.2013  CA 2013 CA.2013.2 Weight  0.942897542
#> 329             Atlantic CA.2011  CA 2011 CA.2011.2 Weight  0.803957279
#> 330             Atlantic CA.2011  CA 2011 CA.2011.1 Weight  0.743956031
#> 331           CO00270-7W CA.2012  CA 2012 CA.2012.2 Weight  0.794595932
#> 332             MSR058-1 CA.2011  CA 2011 CA.2011.1 Weight  0.927006619
#> 333            MSR128-4Y CA.2011  CA 2011 CA.2011.1 Weight  0.742004197
#> 334                NY148 CA.2011  CA 2011 CA.2011.2 Weight  0.563513496
#> 335       NY152(NYH15-5) CA.2013  CA 2013 CA.2013.1 Weight  0.548594198
#> 336              W8603-1 CA.2013  CA 2013 CA.2013.2 Weight  0.662051979
#> 337 Accumulator(W2324-1) CA.2012  CA 2012 CA.2012.2 Weight  0.969064668
#> 338             MSQ089-1 CA.2011  CA 2011 CA.2011.1 Weight  1.075121195
#> 339             MSR058-1 CA.2011  CA 2011 CA.2011.2 Weight  1.057640874
#> 340             MSR127-2 CA.2012  CA 2012 CA.2012.1 Weight  1.156882774
#> 341              W8603-1 CA.2011  CA 2011 CA.2011.2 Weight  0.731064690
#> 342            A01143-3C CA.2013  CA 2013 CA.2013.2 Weight  1.445208429
#> 343             AF4157-6 CA.2011  CA 2011 CA.2011.2 Weight  1.089790016
#> 344           CO02024-9W CA.2011  CA 2011 CA.2011.1 Weight  1.275715985
#> 345             MSR148-4 CA.2011  CA 2011 CA.2011.2 Weight  1.299792377
#> 346             W8539-2Y CA.2011  CA 2011 CA.2011.2 Weight  0.900243680
#> 347              W8615-5 CA.2011  CA 2011 CA.2011.2 Weight  1.050795807
#> 348           AC01151-5W CA.2011  CA 2011 CA.2011.2 Weight  1.527842293
#> 349 Accumulator(W2324-1) CA.2012  CA 2012 CA.2012.1 Weight  1.436841076
#> 350           CO02024-9W CA.2011  CA 2011 CA.2011.2 Weight  1.167324315
#> 351             MSM246-B CA.2011  CA 2011 CA.2011.2 Weight  0.951697157
#> 352              W8603-1 CA.2011  CA 2011 CA.2011.1 Weight  1.370161990
#> 353              W8822-2 CA.2012  CA 2012 CA.2012.1 Weight  1.360552015
#> 354           CO00270-7W CA.2011  CA 2011 CA.2011.2 Weight  1.782248836
#> 355           CO00270-7W CA.2011  CA 2011 CA.2011.1 Weight  1.606139445
#> 356             MSQ089-1 CA.2011  CA 2011 CA.2011.2 Weight  1.528815926
#> 357            MSS165-2Y CA.2011  CA 2011 CA.2011.2 Weight  1.209406380
#> 358             W8539-2Y CA.2011  CA 2011 CA.2011.1 Weight  1.821267867
#> 359             MSR127-2 CA.2011  CA 2011 CA.2011.2 Weight  1.581805839
#> 360            MSS165-2Y CA.2011  CA 2011 CA.2011.1 Weight  1.538216948
#> 361             MSR127-2 CA.2011  CA 2011 CA.2011.1 Weight  2.192971226
#> 362              W8615-5 CA.2011  CA 2011 CA.2011.1 Weight  1.906572817
#> 363            A01143-3C CA.2011  CA 2011 CA.2011.1 Weight  2.633592864
#> 364 Accumulator(W2324-1) CA.2011  CA 2011 CA.2011.1 Weight  1.923875436
#> 365            A01143-3C CA.2011  CA 2011 CA.2011.2 Weight  2.307183443
#> 366            A01143-3C CA.2013  CA 2013 CA.2013.1 Weight  2.718682078
#> 367           AC01151-5W CA.2011  CA 2011 CA.2011.1 Weight  2.040620572
#> 368             MSR148-4 CA.2011  CA 2011 CA.2011.1 Weight  2.621263670
#> 369              W8615-5 CA.2012  CA 2012 CA.2012.2 Weight  3.101811011
#> 370 Accumulator(W2324-1) CA.2011  CA 2011 CA.2011.2 Weight  3.610763025
#>           valueS
#> 1   -1.720715415
#> 2   -1.499497659
#> 3   -1.499497659
#> 4   -1.499497659
#> 5   -1.499497659
#> 6   -1.278279902
#> 7   -1.278279902
#> 8   -1.278279902
#> 9   -1.278279902
#> 10  -1.278279902
#> 11  -1.278279902
#> 12  -1.278279902
#> 13  -1.057062145
#> 14  -1.057062145
#> 15  -1.057062145
#> 16  -1.057062145
#> 17  -1.057062145
#> 18  -1.057062145
#> 19  -1.057062145
#> 20  -1.057062145
#> 21  -1.057062145
#> 22  -1.057062145
#> 23  -1.057062145
#> 24  -1.057062145
#> 25  -1.057062145
#> 26  -1.057062145
#> 27  -1.057062145
#> 28  -1.057062145
#> 29  -1.057062145
#> 30  -1.057062145
#> 31  -0.835844389
#> 32  -0.835844389
#> 33  -0.835844389
#> 34  -0.835844389
#> 35  -0.835844389
#> 36  -0.835844389
#> 37  -0.835844389
#> 38  -0.835844389
#> 39  -0.835844389
#> 40  -0.835844389
#> 41  -0.835844389
#> 42  -0.835844389
#> 43  -0.835844389
#> 44  -0.835844389
#> 45  -0.835844389
#> 46  -0.835844389
#> 47  -0.835844389
#> 48  -0.835844389
#> 49  -0.614626632
#> 50  -0.614626632
#> 51  -0.614626632
#> 52  -0.614626632
#> 53  -0.614626632
#> 54  -0.614626632
#> 55  -0.614626632
#> 56  -0.614626632
#> 57  -0.614626632
#> 58  -0.614626632
#> 59  -0.614626632
#> 60  -0.614626632
#> 61  -0.614626632
#> 62  -0.614626632
#> 63  -0.614626632
#> 64  -0.614626632
#> 65  -0.614626632
#> 66  -0.614626632
#> 67  -0.614626632
#> 68  -0.614626632
#> 69  -0.393408875
#> 70  -0.393408875
#> 71  -0.393408875
#> 72  -0.393408875
#> 73  -0.393408875
#> 74  -0.393408875
#> 75  -0.393408875
#> 76  -0.393408875
#> 77  -0.393408875
#> 78  -0.393408875
#> 79  -0.393408875
#> 80  -0.393408875
#> 81  -0.393408875
#> 82  -0.393408875
#> 83  -0.393408875
#> 84  -0.172191119
#> 85  -0.172191119
#> 86  -0.172191119
#> 87  -0.172191119
#> 88  -0.172191119
#> 89  -0.172191119
#> 90  -0.172191119
#> 91  -0.172191119
#> 92  -0.172191119
#> 93  -0.172191119
#> 94  -0.172191119
#> 95  -0.172191119
#> 96  -0.172191119
#> 97  -0.172191119
#> 98  -0.172191119
#> 99  -0.172191119
#> 100 -0.172191119
#> 101 -0.172191119
#> 102 -0.172191119
#> 103 -0.172191119
#> 104 -0.172191119
#> 105  0.049026638
#> 106  0.049026638
#> 107  0.049026638
#> 108  0.049026638
#> 109  0.049026638
#> 110  0.049026638
#> 111  0.049026638
#> 112  0.049026638
#> 113  0.049026638
#> 114  0.049026638
#> 115  0.049026638
#> 116  0.049026638
#> 117  0.049026638
#> 118  0.049026638
#> 119  0.270244395
#> 120  0.270244395
#> 121  0.270244395
#> 122  0.270244395
#> 123  0.270244395
#> 124  0.270244395
#> 125  0.270244395
#> 126  0.270244395
#> 127  0.270244395
#> 128  0.270244395
#> 129  0.270244395
#> 130  0.270244395
#> 131  0.270244395
#> 132  0.270244395
#> 133  0.491462151
#> 134  0.491462151
#> 135  0.491462151
#> 136  0.491462151
#> 137  0.491462151
#> 138  0.491462151
#> 139  0.491462151
#> 140  0.491462151
#> 141  0.491462151
#> 142  0.491462151
#> 143  0.712679908
#> 144  0.712679908
#> 145  0.712679908
#> 146  0.712679908
#> 147  0.712679908
#> 148  0.712679908
#> 149  0.712679908
#> 150  0.712679908
#> 151  0.712679908
#> 152  0.933897665
#> 153  0.933897665
#> 154  0.933897665
#> 155  0.933897665
#> 156  0.933897665
#> 157  1.155115421
#> 158  1.155115421
#> 159  1.155115421
#> 160  1.155115421
#> 161  1.155115421
#> 162  1.155115421
#> 163  1.376333178
#> 164  1.376333178
#> 165  1.376333178
#> 166  1.376333178
#> 167  1.376333178
#> 168  1.376333178
#> 169  1.597550934
#> 170  1.597550934
#> 171  1.597550934
#> 172  1.597550934
#> 173  1.597550934
#> 174  1.818768691
#> 175  1.818768691
#> 176  2.039986448
#> 177  2.039986448
#> 178  2.261204204
#> 179  2.261204204
#> 180  2.482421961
#> 181  2.482421961
#> 182  2.482421961
#> 183  2.482421961
#> 184  2.924857474
#> 185  3.809728501
#> 186 -1.904710944
#> 187 -1.446958222
#> 188 -1.516271359
#> 189 -1.435510310
#> 190 -1.469051308
#> 191 -1.307167384
#> 192 -1.252885127
#> 193 -1.438962493
#> 194 -1.362001169
#> 195 -1.180293079
#> 196 -0.686940586
#> 197 -0.972147457
#> 198 -1.076444843
#> 199 -1.159478980
#> 200 -1.188148689
#> 201 -1.246607923
#> 202 -1.145567730
#> 203 -0.819852994
#> 204 -1.251570337
#> 205 -1.219755634
#> 206 -1.017792171
#> 207 -0.671175400
#> 208 -1.111328466
#> 209 -1.283317496
#> 210 -1.020208579
#> 211 -1.081052481
#> 212 -0.867852941
#> 213 -1.234828942
#> 214 -0.429062666
#> 215 -1.195027706
#> 216 -0.589984373
#> 217 -0.323892485
#> 218 -1.226764427
#> 219 -0.655897507
#> 220 -0.587283518
#> 221 -0.636425366
#> 222 -0.879965476
#> 223 -0.448960197
#> 224 -0.659728628
#> 225 -0.958733815
#> 226 -0.122812023
#> 227 -1.043605282
#> 228 -0.828186933
#> 229 -0.854903222
#> 230 -0.318770439
#> 231 -0.488842618
#> 232 -0.735799397
#> 233 -0.865993142
#> 234 -0.628242662
#> 235 -0.936532122
#> 236 -0.594419352
#> 237 -0.247575971
#> 238 -0.798393744
#> 239 -0.788407168
#> 240 -0.705221520
#> 241 -0.348264860
#> 242 -0.338483315
#> 243 -0.755715811
#> 244 -0.752285772
#> 245 -0.735646645
#> 246 -0.842713637
#> 247 -0.697167716
#> 248 -0.955954922
#> 249 -0.634630752
#> 250 -1.031659120
#> 251 -0.522481207
#> 252 -1.062704793
#> 253 -0.746255681
#> 254 -0.546320423
#> 255 -0.116574612
#> 256 -0.671755443
#> 257 -0.338181708
#> 258 -0.412168564
#> 259 -0.495521488
#> 260 -0.057343421
#> 261 -0.248603108
#> 262 -0.190226553
#> 263 -0.538487643
#> 264 -0.264204088
#> 265 -0.463717649
#> 266 -0.833799536
#> 267 -0.103570522
#> 268 -0.568484395
#> 269  0.098566285
#> 270  0.066715229
#> 271 -0.524833610
#> 272 -0.325822019
#> 273  0.011424613
#> 274 -0.219217660
#> 275 -0.039319589
#> 276  0.003720692
#> 277  0.022758094
#> 278 -0.202947472
#> 279  0.061383574
#> 280 -0.136667734
#> 281 -0.231594448
#> 282 -0.136450583
#> 283  0.026820367
#> 284 -0.013424286
#> 285 -0.032782056
#> 286 -0.362160719
#> 287 -0.330839989
#> 288  0.018349694
#> 289 -0.420701351
#> 290  0.192229417
#> 291  0.009404862
#> 292  0.056437997
#> 293 -0.248137458
#> 294  0.306390867
#> 295  0.161373012
#> 296  0.011815822
#> 297  0.170630337
#> 298 -0.374530775
#> 299 -0.116689464
#> 300  0.234711307
#> 301  0.144312781
#> 302 -0.072281390
#> 303  0.139732621
#> 304  0.052357810
#> 305  0.152735425
#> 306  0.233158452
#> 307  0.295743756
#> 308  0.181131304
#> 309  0.014473805
#> 310 -0.123372067
#> 311 -0.195697581
#> 312  0.078407572
#> 313  0.306855809
#> 314  0.095513681
#> 315  0.323058995
#> 316  0.300202359
#> 317  0.506922239
#> 318  0.541638035
#> 319  0.322587942
#> 320  0.229329439
#> 321  0.187346766
#> 322  0.058886587
#> 323  0.452509051
#> 324  0.733302892
#> 325  0.830753384
#> 326  0.755462905
#> 327  0.188998663
#> 328  0.942897542
#> 329  0.803957279
#> 330  0.743956031
#> 331  0.794595932
#> 332  0.927006619
#> 333  0.742004197
#> 334  0.563513496
#> 335  0.548594198
#> 336  0.662051979
#> 337  0.969064668
#> 338  1.075121195
#> 339  1.057640874
#> 340  1.156882774
#> 341  0.731064690
#> 342  1.445208429
#> 343  1.089790016
#> 344  1.275715985
#> 345  1.299792377
#> 346  0.900243680
#> 347  1.050795807
#> 348  1.527842293
#> 349  1.436841076
#> 350  1.167324315
#> 351  0.951697157
#> 352  1.370161990
#> 353  1.360552015
#> 354  1.782248836
#> 355  1.606139445
#> 356  1.528815926
#> 357  1.209406380
#> 358  1.821267867
#> 359  1.581805839
#> 360  1.538216948
#> 361  2.192971226
#> 362  1.906572817
#> 363  2.633592864
#> 364  1.923875436
#> 365  2.307183443
#> 366  2.718682078
#> 367  2.040620572
#> 368  2.621263670
#> 369  3.101811011
#> 370  3.610763025
#> 
#> $varG
#>            Yield   Weight
#> Yield  20.434313 4.406246
#> Weight  4.406246 1.000000
#> 
#> $mu
#>        Yield       Weight 
#> 1.177838e+01 9.763680e-18 
#>