variance structure specification

vsr is the main function to build the variance-covariance structure for the random effects to be fitted in the mmer solver.

Usage

vsr(..., Gu=NULL, Gti=NULL, Gtc=NULL, reorderGu=TRUE, buildGu=TRUE)

Arguments

...

variance structure to be specified following the logic desired in the internal kronecker product. For example, if user wants to define a diagonal variance structure for the random effect 'genotypes'(g) with respect to a random effect 'environments'(e), this is:

var(g) = G.e @ I.g

being G.e a matrix containing the variance covariance components for g (genotypes) in each level of e (environments), I.g is the covariance among levels of g (genotypes; i.e. relationship matrix), and @ is the kronecker product. This would be specified in the mmer solver as:

random=~vsr(dsr(e),g)

One strength of sommer is the ability to specify very complex structures with as many kronecker products as desired. For example:

var(g) = G.e @ G.f @ G.h @ I.g

is equivalent to

random=~vsr(e,f,h,g)

where different covariance structures can be applied to the levels of e,f,h (i.e. dsr, usr, csr, atr or a combination of these). For more examples please see the vignettes 'sommer.start' available in the package.

Gu

matrix with the known variance-covariance values for the levels of the u.th random effect (i.e. relationship matrix among individuals or any other known covariance matrix). If NULL, then an identity matrix is assumed. The Gu matrix can have more levels than the ones present in the random effect linked to it but not the other way around. Otherwise, an error message of missing level in Gu will be returned.

Gti

matrix with dimensions t x t (t equal to number of traits) with initial values of the variance-covariance components for the random effect specified in the .... argument. If NULL the program will provide the initial values. The values need to be scaled, see Details section.

Gtc

matrix with dimensions t x t (t equal to number of traits) of constraints for the variance-covariance components for the random effect specified in the ... argument according to the following rules:

0: not to be estimated

1: estimated and constrained to be positive (i.e. variance component)

2: estimated and unconstrained (can be negative or positive, i.e. covariance component)

3: not to be estimated but fixed (value has to be provided in the Gti argument)

In the multi-response scenario if the user doesn't specify this argument the default is to build an unstructured matrix (using the unsm() function). This argument needs to be used wisely since some covariance among responses may not make sense. Useful functions to specify constraints are; diag(), unsm(), fixm().

reorderGu

a TRUE/FALSE statement if the Gu matrix should be reordered based on the names of the design matrix of the random effect or passed with the custom order of the user. This may be important when fitting covariance components in a customized fashion. Only for advanced users.

buildGu

a TRUE/FALSE statement to indicate if the Gu matrix should be built in R when the value for the argument Gu=NULL. Repeat, only when when the value for the argument Gu is equal to NULL. In some cases when the incidence matrix is wide (e.g. rrBLUP models) the covariance structure is a huge p x p matrix that can be avoided when performing matrix operations. By setting this argument to FALSE it allows to skip forming this covariance matrix.

Details

When providing initial values in the Gti argument the user has to provide scaled variance component values. The user can provide values from a previous model by accessing the sigma_scaled output from an mmer model or if an specific value is desired the user can obtain the scaled value as:

m = x/var(y)

where x is the desired initial value and y is the response variable. You can find an example in the DT_cpdata dataset.

Value

$res

a list with all neccesary elements (incidence matrices, known var-cov structures, unknown covariance structures to be estimated and constraints) to be used in the mmer solver.

References

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744

Covarrubias-Pazaran G (2018) Software update: Moving the R package sommer to multivariate mixed models for genome-assisted prediction. doi: https://doi.org/10.1101/354639

Author

Giovanny Covarrubias-Pazaran

See also

Examples

data(DT_example)
DT <- DT_example
A <- A_example

## ============================ ##
## example to without structure
## ============================ ##
dsr(DT$Year)
#> $Z
#>     2011 2012 2013
#> 1      0    0    1
#> 2      0    0    1
#> 3      0    0    1
#> 4      1    0    0
#> 5      0    0    1
#> 6      0    0    1
#> 7      0    1    0
#> 8      1    0    0
#> 9      0    1    0
#> 10     0    1    0
#> 11     0    1    0
#> 12     0    0    1
#> 13     0    0    1
#> 14     0    0    1
#> 15     0    0    1
#> 16     0    1    0
#> 17     0    0    1
#> 18     0    0    1
#> 19     0    0    1
#> 20     0    0    1
#> 21     0    1    0
#> 22     0    1    0
#> 23     0    0    1
#> 24     0    1    0
#> 25     0    0    1
#> 26     0    0    1
#> 27     0    0    1
#> 28     0    0    1
#> 29     0    1    0
#> 30     0    1    0
#> 31     0    1    0
#> 32     0    0    1
#> 33     0    0    1
#> 34     0    1    0
#> 35     0    1    0
#> 36     0    0    1
#> 37     0    0    1
#> 38     0    1    0
#> 39     0    1    0
#> 40     0    1    0
#> 41     0    0    1
#> 42     0    1    0
#> 43     0    1    0
#> 44     0    0    1
#> 45     0    1    0
#> 46     0    1    0
#> 47     0    0    1
#> 48     0    1    0
#> 49     0    1    0
#> 50     0    0    1
#> 51     0    1    0
#> 52     0    0    1
#> 53     1    0    0
#> 54     0    1    0
#> 55     0    0    1
#> 56     0    1    0
#> 57     0    1    0
#> 58     0    0    1
#> 59     0    0    1
#> 60     0    0    1
#> 61     0    1    0
#> 62     0    1    0
#> 63     0    1    0
#> 64     0    1    0
#> 65     0    0    1
#> 66     0    1    0
#> 67     0    0    1
#> 68     0    0    1
#> 69     0    1    0
#> 70     0    1    0
#> 71     0    1    0
#> 72     0    1    0
#> 73     0    1    0
#> 74     0    0    1
#> 75     0    1    0
#> 76     0    1    0
#> 77     1    0    0
#> 78     0    0    1
#> 79     0    0    1
#> 80     0    1    0
#> 81     0    0    1
#> 82     0    1    0
#> 83     0    1    0
#> 84     0    1    0
#> 85     0    1    0
#> 86     0    0    1
#> 87     0    1    0
#> 88     0    0    1
#> 89     0    1    0
#> 90     0    0    1
#> 91     0    1    0
#> 92     0    1    0
#> 93     0    1    0
#> 94     0    0    1
#> 95     0    1    0
#> 96     0    0    1
#> 97     0    0    1
#> 98     1    0    0
#> 99     0    1    0
#> 100    1    0    0
#> 101    0    1    0
#> 102    0    0    1
#> 103    0    0    1
#> 104    0    0    1
#> 105    0    0    1
#> 106    0    1    0
#> 107    0    0    1
#> 108    0    1    0
#> 109    0    1    0
#> 110    0    1    0
#> 111    0    0    1
#> 112    1    0    0
#> 113    0    1    0
#> 114    0    1    0
#> 115    0    1    0
#> 116    0    0    1
#> 117    0    1    0
#> 118    0    1    0
#> 119    1    0    0
#> 120    1    0    0
#> 121    0    0    1
#> 122    0    1    0
#> 123    1    0    0
#> 124    0    1    0
#> 125    0    1    0
#> 126    0    1    0
#> 127    0    1    0
#> 128    0    1    0
#> 129    0    0    1
#> 130    0    0    1
#> 131    0    1    0
#> 132    0    0    1
#> 133    1    0    0
#> 134    0    1    0
#> 135    0    1    0
#> 136    0    0    1
#> 137    0    0    1
#> 138    0    0    1
#> 139    0    1    0
#> 140    0    0    1
#> 141    0    0    1
#> 142    0    1    0
#> 143    0    0    1
#> 144    1    0    0
#> 145    1    0    0
#> 146    0    1    0
#> 147    1    0    0
#> 148    1    0    0
#> 149    1    0    0
#> 150    0    0    1
#> 151    0    0    1
#> 152    0    1    0
#> 153    1    0    0
#> 154    1    0    0
#> 155    0    1    0
#> 156    1    0    0
#> 157    0    0    1
#> 158    1    0    0
#> 159    1    0    0
#> 160    1    0    0
#> 161    1    0    0
#> 162    1    0    0
#> 163    1    0    0
#> 164    0    1    0
#> 165    1    0    0
#> 166    1    0    0
#> 167    1    0    0
#> 168    0    1    0
#> 169    1    0    0
#> 170    1    0    0
#> 171    1    0    0
#> 172    1    0    0
#> 173    1    0    0
#> 174    1    0    0
#> 175    1    0    0
#> 176    1    0    0
#> 177    1    0    0
#> 178    1    0    0
#> 179    1    0    0
#> 180    1    0    0
#> 181    0    0    1
#> 182    1    0    0
#> 183    1    0    0
#> 184    0    1    0
#> 185    1    0    0
#> attr(,"assign")
#> [1] 1 1 1
#> attr(,"contrasts")
#> attr(,"contrasts")$dummy
#> [1] "contr.treatment"
#> 
#> 
#> $thetaC
#>      2011 2012 2013
#> 2011    1    0    0
#> 2012    0    1    0
#> 2013    0    0    1
#> 
mix <- mmer(Yield~Env, 
            random= ~ vsr(Name),
            rcov=~ vsr(units),
            data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -36.8096   1:17:27      0           0
#>     2      -33.211   1:17:27      0           0
#>     3      -32.6234   1:17:27      0           0
#>     4      -32.5942   1:17:27      0           0
#>     5      -32.5942   1:17:27      0           0
            
## ============================ ##
## example to without structure but 
## using covariance among levels in the
## random effect Name
## ============================ ##
dsr(DT$Year)
#> $Z
#>     2011 2012 2013
#> 1      0    0    1
#> 2      0    0    1
#> 3      0    0    1
#> 4      1    0    0
#> 5      0    0    1
#> 6      0    0    1
#> 7      0    1    0
#> 8      1    0    0
#> 9      0    1    0
#> 10     0    1    0
#> 11     0    1    0
#> 12     0    0    1
#> 13     0    0    1
#> 14     0    0    1
#> 15     0    0    1
#> 16     0    1    0
#> 17     0    0    1
#> 18     0    0    1
#> 19     0    0    1
#> 20     0    0    1
#> 21     0    1    0
#> 22     0    1    0
#> 23     0    0    1
#> 24     0    1    0
#> 25     0    0    1
#> 26     0    0    1
#> 27     0    0    1
#> 28     0    0    1
#> 29     0    1    0
#> 30     0    1    0
#> 31     0    1    0
#> 32     0    0    1
#> 33     0    0    1
#> 34     0    1    0
#> 35     0    1    0
#> 36     0    0    1
#> 37     0    0    1
#> 38     0    1    0
#> 39     0    1    0
#> 40     0    1    0
#> 41     0    0    1
#> 42     0    1    0
#> 43     0    1    0
#> 44     0    0    1
#> 45     0    1    0
#> 46     0    1    0
#> 47     0    0    1
#> 48     0    1    0
#> 49     0    1    0
#> 50     0    0    1
#> 51     0    1    0
#> 52     0    0    1
#> 53     1    0    0
#> 54     0    1    0
#> 55     0    0    1
#> 56     0    1    0
#> 57     0    1    0
#> 58     0    0    1
#> 59     0    0    1
#> 60     0    0    1
#> 61     0    1    0
#> 62     0    1    0
#> 63     0    1    0
#> 64     0    1    0
#> 65     0    0    1
#> 66     0    1    0
#> 67     0    0    1
#> 68     0    0    1
#> 69     0    1    0
#> 70     0    1    0
#> 71     0    1    0
#> 72     0    1    0
#> 73     0    1    0
#> 74     0    0    1
#> 75     0    1    0
#> 76     0    1    0
#> 77     1    0    0
#> 78     0    0    1
#> 79     0    0    1
#> 80     0    1    0
#> 81     0    0    1
#> 82     0    1    0
#> 83     0    1    0
#> 84     0    1    0
#> 85     0    1    0
#> 86     0    0    1
#> 87     0    1    0
#> 88     0    0    1
#> 89     0    1    0
#> 90     0    0    1
#> 91     0    1    0
#> 92     0    1    0
#> 93     0    1    0
#> 94     0    0    1
#> 95     0    1    0
#> 96     0    0    1
#> 97     0    0    1
#> 98     1    0    0
#> 99     0    1    0
#> 100    1    0    0
#> 101    0    1    0
#> 102    0    0    1
#> 103    0    0    1
#> 104    0    0    1
#> 105    0    0    1
#> 106    0    1    0
#> 107    0    0    1
#> 108    0    1    0
#> 109    0    1    0
#> 110    0    1    0
#> 111    0    0    1
#> 112    1    0    0
#> 113    0    1    0
#> 114    0    1    0
#> 115    0    1    0
#> 116    0    0    1
#> 117    0    1    0
#> 118    0    1    0
#> 119    1    0    0
#> 120    1    0    0
#> 121    0    0    1
#> 122    0    1    0
#> 123    1    0    0
#> 124    0    1    0
#> 125    0    1    0
#> 126    0    1    0
#> 127    0    1    0
#> 128    0    1    0
#> 129    0    0    1
#> 130    0    0    1
#> 131    0    1    0
#> 132    0    0    1
#> 133    1    0    0
#> 134    0    1    0
#> 135    0    1    0
#> 136    0    0    1
#> 137    0    0    1
#> 138    0    0    1
#> 139    0    1    0
#> 140    0    0    1
#> 141    0    0    1
#> 142    0    1    0
#> 143    0    0    1
#> 144    1    0    0
#> 145    1    0    0
#> 146    0    1    0
#> 147    1    0    0
#> 148    1    0    0
#> 149    1    0    0
#> 150    0    0    1
#> 151    0    0    1
#> 152    0    1    0
#> 153    1    0    0
#> 154    1    0    0
#> 155    0    1    0
#> 156    1    0    0
#> 157    0    0    1
#> 158    1    0    0
#> 159    1    0    0
#> 160    1    0    0
#> 161    1    0    0
#> 162    1    0    0
#> 163    1    0    0
#> 164    0    1    0
#> 165    1    0    0
#> 166    1    0    0
#> 167    1    0    0
#> 168    0    1    0
#> 169    1    0    0
#> 170    1    0    0
#> 171    1    0    0
#> 172    1    0    0
#> 173    1    0    0
#> 174    1    0    0
#> 175    1    0    0
#> 176    1    0    0
#> 177    1    0    0
#> 178    1    0    0
#> 179    1    0    0
#> 180    1    0    0
#> 181    0    0    1
#> 182    1    0    0
#> 183    1    0    0
#> 184    0    1    0
#> 185    1    0    0
#> attr(,"assign")
#> [1] 1 1 1
#> attr(,"contrasts")
#> attr(,"contrasts")$dummy
#> [1] "contr.treatment"
#> 
#> 
#> $thetaC
#>      2011 2012 2013
#> 2011    1    0    0
#> 2012    0    1    0
#> 2013    0    0    1
#> 
mix <- mmer(Yield~Env, 
            random= ~ vsr(Name, Gu=A),
            rcov=~ vsr(units),
            data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -36.8096   1:17:27      0           0
#>     2      -33.211   1:17:27      0           0
#>     3      -32.6234   1:17:27      0           0
#>     4      -32.5942   1:17:27      0           0
#>     5      -32.5942   1:17:27      0           0
            
## ============================ ##
## example to use dsr() structure (DIAGONAL)
## ============================ ##
dsr(DT$Year)
#> $Z
#>     2011 2012 2013
#> 1      0    0    1
#> 2      0    0    1
#> 3      0    0    1
#> 4      1    0    0
#> 5      0    0    1
#> 6      0    0    1
#> 7      0    1    0
#> 8      1    0    0
#> 9      0    1    0
#> 10     0    1    0
#> 11     0    1    0
#> 12     0    0    1
#> 13     0    0    1
#> 14     0    0    1
#> 15     0    0    1
#> 16     0    1    0
#> 17     0    0    1
#> 18     0    0    1
#> 19     0    0    1
#> 20     0    0    1
#> 21     0    1    0
#> 22     0    1    0
#> 23     0    0    1
#> 24     0    1    0
#> 25     0    0    1
#> 26     0    0    1
#> 27     0    0    1
#> 28     0    0    1
#> 29     0    1    0
#> 30     0    1    0
#> 31     0    1    0
#> 32     0    0    1
#> 33     0    0    1
#> 34     0    1    0
#> 35     0    1    0
#> 36     0    0    1
#> 37     0    0    1
#> 38     0    1    0
#> 39     0    1    0
#> 40     0    1    0
#> 41     0    0    1
#> 42     0    1    0
#> 43     0    1    0
#> 44     0    0    1
#> 45     0    1    0
#> 46     0    1    0
#> 47     0    0    1
#> 48     0    1    0
#> 49     0    1    0
#> 50     0    0    1
#> 51     0    1    0
#> 52     0    0    1
#> 53     1    0    0
#> 54     0    1    0
#> 55     0    0    1
#> 56     0    1    0
#> 57     0    1    0
#> 58     0    0    1
#> 59     0    0    1
#> 60     0    0    1
#> 61     0    1    0
#> 62     0    1    0
#> 63     0    1    0
#> 64     0    1    0
#> 65     0    0    1
#> 66     0    1    0
#> 67     0    0    1
#> 68     0    0    1
#> 69     0    1    0
#> 70     0    1    0
#> 71     0    1    0
#> 72     0    1    0
#> 73     0    1    0
#> 74     0    0    1
#> 75     0    1    0
#> 76     0    1    0
#> 77     1    0    0
#> 78     0    0    1
#> 79     0    0    1
#> 80     0    1    0
#> 81     0    0    1
#> 82     0    1    0
#> 83     0    1    0
#> 84     0    1    0
#> 85     0    1    0
#> 86     0    0    1
#> 87     0    1    0
#> 88     0    0    1
#> 89     0    1    0
#> 90     0    0    1
#> 91     0    1    0
#> 92     0    1    0
#> 93     0    1    0
#> 94     0    0    1
#> 95     0    1    0
#> 96     0    0    1
#> 97     0    0    1
#> 98     1    0    0
#> 99     0    1    0
#> 100    1    0    0
#> 101    0    1    0
#> 102    0    0    1
#> 103    0    0    1
#> 104    0    0    1
#> 105    0    0    1
#> 106    0    1    0
#> 107    0    0    1
#> 108    0    1    0
#> 109    0    1    0
#> 110    0    1    0
#> 111    0    0    1
#> 112    1    0    0
#> 113    0    1    0
#> 114    0    1    0
#> 115    0    1    0
#> 116    0    0    1
#> 117    0    1    0
#> 118    0    1    0
#> 119    1    0    0
#> 120    1    0    0
#> 121    0    0    1
#> 122    0    1    0
#> 123    1    0    0
#> 124    0    1    0
#> 125    0    1    0
#> 126    0    1    0
#> 127    0    1    0
#> 128    0    1    0
#> 129    0    0    1
#> 130    0    0    1
#> 131    0    1    0
#> 132    0    0    1
#> 133    1    0    0
#> 134    0    1    0
#> 135    0    1    0
#> 136    0    0    1
#> 137    0    0    1
#> 138    0    0    1
#> 139    0    1    0
#> 140    0    0    1
#> 141    0    0    1
#> 142    0    1    0
#> 143    0    0    1
#> 144    1    0    0
#> 145    1    0    0
#> 146    0    1    0
#> 147    1    0    0
#> 148    1    0    0
#> 149    1    0    0
#> 150    0    0    1
#> 151    0    0    1
#> 152    0    1    0
#> 153    1    0    0
#> 154    1    0    0
#> 155    0    1    0
#> 156    1    0    0
#> 157    0    0    1
#> 158    1    0    0
#> 159    1    0    0
#> 160    1    0    0
#> 161    1    0    0
#> 162    1    0    0
#> 163    1    0    0
#> 164    0    1    0
#> 165    1    0    0
#> 166    1    0    0
#> 167    1    0    0
#> 168    0    1    0
#> 169    1    0    0
#> 170    1    0    0
#> 171    1    0    0
#> 172    1    0    0
#> 173    1    0    0
#> 174    1    0    0
#> 175    1    0    0
#> 176    1    0    0
#> 177    1    0    0
#> 178    1    0    0
#> 179    1    0    0
#> 180    1    0    0
#> 181    0    0    1
#> 182    1    0    0
#> 183    1    0    0
#> 184    0    1    0
#> 185    1    0    0
#> attr(,"assign")
#> [1] 1 1 1
#> attr(,"contrasts")
#> attr(,"contrasts")$dummy
#> [1] "contr.treatment"
#> 
#> 
#> $thetaC
#>      2011 2012 2013
#> 2011    1    0    0
#> 2012    0    1    0
#> 2013    0    0    1
#> 
mix <- mmer(Yield~Env, 
            random= ~ vsr(dsr(Year),Name),
            rcov=~ vsr(dsr(Year),units),
            data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -42.26   1:17:27      0           0
#>     2      -25.3744   1:17:27      0           0
#>     3      -19.1877   1:17:27      0           0
#>     4      -18.3538   1:17:27      0           0
#>     5      -18.3432   1:17:27      0           0
#>     6      -18.343   1:17:27      0           0

## ============================ ##
## example to use atr() structure (level-specific)
## ============================ ##
unique(DT$Year)
#> [1] 2013 2011 2012
#> Levels: 2011 2012 2013
mix <- mmer(Yield~Env, 
            random= ~ vsr(atr(Year,c("2011","2012")),Name),
            rcov=~ vsr(dsr(Year),units),
            data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -45.5289   1:17:27      0           0
#>     2      -35.9052   1:17:27      0           0
#>     3      -33.3779   1:17:27      0           0
#>     4      -33.0963   1:17:27      0           0
#>     5      -33.0911   1:17:27      0           0
#>     6      -33.091   1:17:27      0           0

## ============================ ##
## example to use usr() structure (UNSTRUCTURED)
## ============================ ##
usr(DT$Year)
#> $Z
#>     2011 2012 2013
#> 1      0    0    1
#> 2      0    0    1
#> 3      0    0    1
#> 4      1    0    0
#> 5      0    0    1
#> 6      0    0    1
#> 7      0    1    0
#> 8      1    0    0
#> 9      0    1    0
#> 10     0    1    0
#> 11     0    1    0
#> 12     0    0    1
#> 13     0    0    1
#> 14     0    0    1
#> 15     0    0    1
#> 16     0    1    0
#> 17     0    0    1
#> 18     0    0    1
#> 19     0    0    1
#> 20     0    0    1
#> 21     0    1    0
#> 22     0    1    0
#> 23     0    0    1
#> 24     0    1    0
#> 25     0    0    1
#> 26     0    0    1
#> 27     0    0    1
#> 28     0    0    1
#> 29     0    1    0
#> 30     0    1    0
#> 31     0    1    0
#> 32     0    0    1
#> 33     0    0    1
#> 34     0    1    0
#> 35     0    1    0
#> 36     0    0    1
#> 37     0    0    1
#> 38     0    1    0
#> 39     0    1    0
#> 40     0    1    0
#> 41     0    0    1
#> 42     0    1    0
#> 43     0    1    0
#> 44     0    0    1
#> 45     0    1    0
#> 46     0    1    0
#> 47     0    0    1
#> 48     0    1    0
#> 49     0    1    0
#> 50     0    0    1
#> 51     0    1    0
#> 52     0    0    1
#> 53     1    0    0
#> 54     0    1    0
#> 55     0    0    1
#> 56     0    1    0
#> 57     0    1    0
#> 58     0    0    1
#> 59     0    0    1
#> 60     0    0    1
#> 61     0    1    0
#> 62     0    1    0
#> 63     0    1    0
#> 64     0    1    0
#> 65     0    0    1
#> 66     0    1    0
#> 67     0    0    1
#> 68     0    0    1
#> 69     0    1    0
#> 70     0    1    0
#> 71     0    1    0
#> 72     0    1    0
#> 73     0    1    0
#> 74     0    0    1
#> 75     0    1    0
#> 76     0    1    0
#> 77     1    0    0
#> 78     0    0    1
#> 79     0    0    1
#> 80     0    1    0
#> 81     0    0    1
#> 82     0    1    0
#> 83     0    1    0
#> 84     0    1    0
#> 85     0    1    0
#> 86     0    0    1
#> 87     0    1    0
#> 88     0    0    1
#> 89     0    1    0
#> 90     0    0    1
#> 91     0    1    0
#> 92     0    1    0
#> 93     0    1    0
#> 94     0    0    1
#> 95     0    1    0
#> 96     0    0    1
#> 97     0    0    1
#> 98     1    0    0
#> 99     0    1    0
#> 100    1    0    0
#> 101    0    1    0
#> 102    0    0    1
#> 103    0    0    1
#> 104    0    0    1
#> 105    0    0    1
#> 106    0    1    0
#> 107    0    0    1
#> 108    0    1    0
#> 109    0    1    0
#> 110    0    1    0
#> 111    0    0    1
#> 112    1    0    0
#> 113    0    1    0
#> 114    0    1    0
#> 115    0    1    0
#> 116    0    0    1
#> 117    0    1    0
#> 118    0    1    0
#> 119    1    0    0
#> 120    1    0    0
#> 121    0    0    1
#> 122    0    1    0
#> 123    1    0    0
#> 124    0    1    0
#> 125    0    1    0
#> 126    0    1    0
#> 127    0    1    0
#> 128    0    1    0
#> 129    0    0    1
#> 130    0    0    1
#> 131    0    1    0
#> 132    0    0    1
#> 133    1    0    0
#> 134    0    1    0
#> 135    0    1    0
#> 136    0    0    1
#> 137    0    0    1
#> 138    0    0    1
#> 139    0    1    0
#> 140    0    0    1
#> 141    0    0    1
#> 142    0    1    0
#> 143    0    0    1
#> 144    1    0    0
#> 145    1    0    0
#> 146    0    1    0
#> 147    1    0    0
#> 148    1    0    0
#> 149    1    0    0
#> 150    0    0    1
#> 151    0    0    1
#> 152    0    1    0
#> 153    1    0    0
#> 154    1    0    0
#> 155    0    1    0
#> 156    1    0    0
#> 157    0    0    1
#> 158    1    0    0
#> 159    1    0    0
#> 160    1    0    0
#> 161    1    0    0
#> 162    1    0    0
#> 163    1    0    0
#> 164    0    1    0
#> 165    1    0    0
#> 166    1    0    0
#> 167    1    0    0
#> 168    0    1    0
#> 169    1    0    0
#> 170    1    0    0
#> 171    1    0    0
#> 172    1    0    0
#> 173    1    0    0
#> 174    1    0    0
#> 175    1    0    0
#> 176    1    0    0
#> 177    1    0    0
#> 178    1    0    0
#> 179    1    0    0
#> 180    1    0    0
#> 181    0    0    1
#> 182    1    0    0
#> 183    1    0    0
#> 184    0    1    0
#> 185    1    0    0
#> attr(,"assign")
#> [1] 1 1 1
#> attr(,"contrasts")
#> attr(,"contrasts")$dummy
#> [1] "contr.treatment"
#> 
#> 
#> $thetaC
#>      2011 2012 2013
#> 2011    1    2    2
#> 2012    2    1    2
#> 2013    2    2    1
#> 
mix <- mmer(Yield~Env, 
            random= ~ vsr(usr(Year),Name),
            rcov=~ vsr(dsr(Year),units),
            data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -37.9059   1:17:28      0           0
#>     2      -17.9745   1:17:28      0           0
#>     3      -12.2427   1:17:28      0           0
#>     4      -11.5121   1:17:28      0           0
#>     5      -11.5001   1:17:28      0           0
#>     6      -11.4997   1:17:28      0           0

## ============================ ##
## example to use csr() structure (CUSTOMIZED)
## ============================ ##
unique(DT$Year)
#> [1] 2013 2011 2012
#> Levels: 2011 2012 2013
mm <- matrix(1,3,3); mm[1,3] <- mm[3,1] <- 0;mm #don't estimate cov 2011-2013
#>      [,1] [,2] [,3]
#> [1,]    1    1    0
#> [2,]    1    1    1
#> [3,]    0    1    1
mix <- mmer(Yield~Env, 
            random= ~ vsr(csr(Year,mm),Name),
            rcov=~ vsr(dsr(Year),units),
            data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -40.2254   1:17:28      0           0
#>     2      -22.056   1:17:28      0           0
#>     3      -15.6387   1:17:28      0           0
#>     4      -14.6846   1:17:28      0           0
#>     5      -14.6138   1:17:28      0           0
#>     6      -14.594   1:17:28      0           0
#>     7      -14.5877   1:17:29      1           0
#>     8      -14.5857   1:17:29      1           0
#>     9      -14.5851   1:17:29      1           0

## ============================ ##
## example to use overlay() + vsr() structure
## ============================ ##
data("DT_halfdiallel")
DT <- DT_halfdiallel
head(DT)
#>   rep geno male female     sugar
#> 1   1   12    1      2 13.950509
#> 2   2   12    1      2  9.756918
#> 3   1   13    1      3 13.906355
#> 4   2   13    1      3  9.119455
#> 5   1   14    1      4  5.174483
#> 6   2   14    1      4  8.452221
DT$femalef <- as.factor(DT$female)
DT$malef <- as.factor(DT$male)
DT$genof <- as.factor(DT$geno)
A <- diag(7); colnames(A) <- rownames(A) <- 1:7;A # if you want to provide a covariance matrix
#>   1 2 3 4 5 6 7
#> 1 1 0 0 0 0 0 0
#> 2 0 1 0 0 0 0 0
#> 3 0 0 1 0 0 0 0
#> 4 0 0 0 1 0 0 0
#> 5 0 0 0 0 1 0 0
#> 6 0 0 0 0 0 1 0
#> 7 0 0 0 0 0 0 1
#### model using overlay
modh <- mmer(sugar~1, 
             random=~vsr(overlay(femalef,malef,sparse=FALSE), Gu=A) 
                     + genof,
              data=DT)
#> iteration    LogLik     wall    cpu(sec)   restrained
#>     1      -10.425   1:17:29      0           0
#>     2      -6.487   1:17:29      0           0
#>     3      -5.732   1:17:29      0           0
#>     4      -5.67494   1:17:29      0           0
#>     5      -5.67441   1:17:29      0           0

## ============================ ##
## example to use vsr() + dsr() + spl2D() structure
## ============================ ##            
# ### mimic two fields
# data(DT_cpdata)
# DT <- DT_cpdata
# GT <- GT_cpdata
# MP <- MP_cpdata
# aa <- DT; bb <- DT
# aa$FIELD <- "A";bb$FIELD <- "B"
# set.seed(1234)
# aa$Yield <- aa$Yield + rnorm(length(aa$Yield),0,4)
# DT2 <- rbind(aa,bb)
# head(DT2)
# 
# mix <- mmer(Yield~1,
#             random=~vsr(dsr(FIELD),id, Gu=A) + 
#               vsr(dsr(FIELD),Rowf) +
#               vsr(dsr(FIELD),Colf) +
#                 vsr(dsr(FIELD),spl2D(Row,Col)),
#             rcov=~vsr(dsr(FIELD),units),
#             data=DT2)